% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{add_latest_eval}
\alias{add_latest_eval}
\title{Add eval time to msrawfiles}
\usage{
add_latest_eval(escon, index, esid, fieldName = "dbas_last_eval")
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{index}{must be an msrawfiles index}

\item{esid}{elasicsearch id doc to be updated}

\item{fieldName}{field name to be updated, must be either dbas_last_eval or dbas_is_last_eval}
}
\value{
true if successful
}
\description{
Add eval time to msrawfiles
}
