% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{add_rawfiles}
\alias{add_rawfiles}
\title{Add new MS measurement files to msrawfiles index}
\usage{
add_rawfiles(
  escon,
  rfindex,
  templateId,
  newPaths,
  newStart = "filename",
  newStation = "same_as_template"
)
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{templateId}{}

\item{newPaths}{}

\item{newStart}{Either "filename" (default), meaning extract the start time 
from the filename or provide a start date as an 8 digit number "YYYYMMDD"}

\item{newStation}{Either be "same_as_template" (default), which will
copy the value from the template document, or "filename", meaning extract
the station and loc values from the code in the filename (will compare to 
other docs in msrawfiles index) or provide a list of length 2 with station 
and loc values. See details.}
}
\description{
Add new MS measurement files to msrawfiles index
}
\details{
newStation can be either "same_as_template" or "filename" or
a new fixed station name and location (list with fields "station" and "loc")
"loc" being a geopoint with fields "lat" and "lon"
if it is in the filename it will use the dbas_station_regex field in
the index to get the station information from another doc in the index
}
