% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{add_sha256_spectral_library}
\alias{add_sha256_spectral_library}
\title{Add hash of spectral library in current state to msrawfiles}
\usage{
add_sha256_spectral_library(escon, rfindex, esid)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{esid}{ElasticSearch document IDs (character)}
}
\value{
Returns TRUE if successful
}
\description{
This is used to see with which spectral library the last evaluation was done with.
}
