% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{change_msrawfile_filename}
\alias{change_msrawfile_filename}
\title{Change Filename in the rawfiles-db an the data-saving structure}
\usage{
change_msrawfile_filename(escon, rfindex, oldName, newName, oldPath, newPath)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{oldName}{originally filename}

\item{newName}{updated filename}

\item{oldPath}{originally path}

\item{newPath}{updated path}
}
\value{
log_info("File updated") if successful
}
\description{
Change Filename in the rawfiles-db an the data-saving structure
}
