% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-modification.R
\name{changeMsrawfileFilename}
\alias{changeMsrawfileFilename}
\title{Change filename in the msrawfiles-db and on the filesystem}
\usage{
changeMsrawfileFilename(rfIndex, oldPath, newPath)
}
\arguments{
\item{rfIndex}{index name for msrawfiles index}

\item{oldPath}{original path}

\item{newPath}{updated path}
}
\value{
True (invisibly) if successful
}
\description{
This function will change both filename and path of the doc.
}
