% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-modification.R
\name{changeMsrawfilePath}
\alias{changeMsrawfilePath}
\title{Change the path of one document in msrawfiles}
\usage{
changeMsrawfilePath(rfIndex, oldPath, newPath, checkType = "md5")
}
\arguments{
\item{rfIndex}{index name for rawfiles index}

\item{oldPath}{Current path in rawfiles index, file must exist (1 file)}

\item{newPath}{New path, file must exist (1 file)}

\item{checkType}{Method to check that files are the same, either "md5" (default) or "filesize"}
}
\value{
TRUE if change was successful (invisibly)
}
\description{
This is a secure method of changing the path of a doc in msrawfiles. 
Filename is not changed and must remain the same.
}
\details{
Both files must exist when making the change. The function takes some time 
if it compares md5-checksums of the two files. For large numbers of files use the
filesize check.
}
\examples{
\dontrun{
library(ntsportal)
connectNtsportal()

ind <- "ntsp_msrawfiles"
res <- esSearchPaged(ind, sort = "path", searchBody = list(query = list(regexp = list(path = "/srv.*"))), source = "path")$hits$hits
partToRemove <- "/srv/cifs-mounts/g2/G/G2/HRMS/Messdaten/"
paths <- sapply(res, function(x) x[["_source"]]$path)
oldPaths <- data.frame(oldPath = paths, relPath = sub(partToRemove, "", paths))
relPaths <- list.files("/beegfs/nts/ntsportal/msrawfiles", recursive = T)
newPaths <- data.frame(
  newPath = paste0("/beegfs/nts/ntsportal/msrawfiles/", relPaths),
  relPath = relPaths
)

oldAndNewPaths <- merge(oldPaths, newPaths, all.x = T, by = "relPath")

for (i in 1:nrow(oldAndNewPaths)) {
  changeMsrawfilePath(ind, oldAndNewPaths$oldPath[i], oldAndNewPaths$newPath[i], checkType = "filesize")  
}
}
}
