% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles_quality_assurance.R
\name{check_batches_eval}
\alias{check_batches_eval}
\title{Check that evaluation batches have consistent settings}
\usage{
check_batches_eval(escon, rfindex, batches)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{Elasticsearch index name for msrawfiles index}

\item{batches}{a list of batches, where each batch is character vector of 
elasticsearch document ids}
}
\value{
TRUE if successful
}
\description{
Evaluation is run in batches and certain settings must be the same
for all files in the batch.
}
