% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{check_integrity_msrawfiles}
\alias{check_integrity_msrawfiles}
\title{Check consistency of msrawfiles DB}
\usage{
check_integrity_msrawfiles(escon, rfindex)
}
\arguments{
\item{escon}{}

\item{rfindex}{}
}
\value{
TRUE if successful (invisibly)
}
\description{
Must be done before any processing operation
}
