% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles_quality_assurance.R
\name{check_integrity_msrawfiles_nts}
\alias{check_integrity_msrawfiles_nts}
\title{Check consistency of msrawfiles for nts data}
\usage{
check_integrity_msrawfiles_nts(msrawfilesDocsList)
}
\arguments{
\item{msrawfilesDocsList}{msrawfiles_docs_list object (from Elasticsearch API)}
}
\value{
TRUE if completed successfully
}
\description{
The function will throw an exception if an error is found. It is designed to
be run after uploading data and before processing.
}
