% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{checkPngAvailability}
\alias{checkPngAvailability}
\title{Check that all structural formula PNGs have been created}
\usage{
checkPngAvailability(databasePath, targetDir)
}
\arguments{
\item{databasePath}{Path to spectral library (CSL, SQLite format)}

\item{targetDir}{Path where the PNGs are to be saved}
}
\description{
Checks that PNGs for structural formulae are 
consistent with the compounds in the spectral library. It
compares the InChI-Keys in the library and filenames and
prints any inconsistencies to the console.
}
