% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{checkPngSyncronization}
\alias{checkPngSyncronization}
\title{Check synchronization with the picture server}
\usage{
checkPngSyncronization(fileUrl, filePathHd)
}
\arguments{
\item{fileUrl}{URL of the file on the picture server}

\item{filePathHd}{Path to the file on disk}
}
\description{
Checks that a PNG file on the picture server and on disk are identical.
}
