% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conn_udb}
\alias{conn_udb}
\title{Connect to a sqlite db (ufid-library)}
\usage{
conn_udb(pth)
}
\arguments{
\item{pth}{}
}
\value{
connection object
}
\description{
Connect to a sqlite db (ufid-library)
}
