% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionToElasticsearch.R
\name{connectNtsportal}
\alias{connectNtsportal}
\title{Set ElasticSearch credentials and create ElasticSearch connection object}
\usage{
connectNtsportal()
}
\description{
If credentials are available, will create the "escon" connection object in the 
global environment for use by all functions. Otherwise will ask for input of
credentials (in interactive use). This function uses the keyring package to 
store and retrieve credentials. For working on Linux in the console, the 'file'
backend must be used, see https://keyring.r-lib.org/reference/backends.html.
This can be set with Sys.setenv("R_KEYRING_BACKEND" = "file")
}
