% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-convertToRecord.R
\name{convertToRecord.dbasResult}
\alias{convertToRecord.dbasResult}
\title{Convert a dbasResult to a list of featureRecords}
\usage{
\method{convertToRecord}{dbasResult}(scanResult, msrawfileRecords)
}
\description{
Convert a dbasResult to a list of featureRecords
}
\examples{
\dontrun{
dbComm <- getDbComm()
recs <- getTableAsRecords(
  dbComm, 
  "ntsp25.1_msrawfiles", 
  searchBlock = list(query = list(regexp = list(filename = "Des_.._.._pos.mzXML"))), 
  newMsrawfilesRecord
)
recsBlanks <- getTableAsRecords(
  dbComm, 
  "ntsp25.1_msrawfiles", 
  searchBlock = list(query = list(regexp = list(path = ".*mud_pos/BW.*"))), 
  newMsrawfilesRecord
)
res <- scanBatchDbas(c(recs, recsBlanks), "Methyltriphenylphosphonium")
featureRecs <- convertToRecord(res, c(recs, recsBlanks))
}

}
