% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-convertToRecord-utils.R,
%   R/screening-dbasConvertToRecord.R, R/screening-ntsConvertToRecord.R
\name{convertToRecord}
\alias{convertToRecord}
\alias{convertToRecord.dbasResult}
\alias{convertToRecord.ntsResult}
\title{Convert the result of file scanning to a list of feature records
Compound and sample metadata, including spectra are collected from the `scanResult`, from the linked raw measurement
files, or from `msrawfileRecords`}
\usage{
convertToRecord(scanResult, msrawfilesBatch)

\method{convertToRecord}{dbasResult}(scanResult, msrawfilesBatch)

\method{convertToRecord}{ntsResult}(scanResult, msrawfilesBatch = NA)
}
\arguments{
\item{scanResult}{results of a file scanning algorithm (e.g. dbasResult, ntsResult)}

\item{msrawfilesBatch}{msrawfilesBatch used to create the scanResult}
}
\value{
a `list` of `featureRecord`s
}
\description{
The results of DBAS file scanning (`dbasResult`) are converted to the NTSPortal `featureRecord` format.
Compound and sample metadata are collected from the `msrawfileRecords` argument and the spectral library (CSL).

The results of NTS file scanning (`ntsResult`) are converted to the NTSPortal `featureRecord` format.
Annotated features (matching a substance in the spectral library) are removed.
}
\details{
This method does not use `msrawfilesBatch` but it needs to be in the method definition for polymorphism, 
e.g. in `screeningOneBatch()`.
}
\examples{
\dontrun{
dbComm <- getDbComm()
recs <- getTableAsRecords(
  dbComm, 
  "ntsp25.3_msrawfiles", 
  searchBlock = list(query = list(regexp = list(filename = "Des_.._.._pos.mzXML"))), 
  newDbasMsrawfilesRecord
)
recsBlanks <- getTableAsRecords(
  dbComm, 
  "ntsp25.3_msrawfiles", 
  searchBlock = list(query = list(regexp = list(path = ".*mud_pos/BW.*"))), 
  newDbasMsrawfilesRecord
)
dbasBatch <- recordsToOneBatch(c(recs, recsBlanks))
dbasRes <- scanBatch(dbasBatch, "Methyltriphenylphosphonium")
featureRecs <- convertToRecord(dbasRes, dbasBatch)
}
\dontrun{
  testMsrawfilesRecords <- getRecordsSampleAndBlank()
  testNtsResults <- scanBatch(testMsrawfilesRecords)
  features <- convertToRecord(testNtsResults)  
}
}
