% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{create_dbas_index}
\alias{create_dbas_index}
\title{Create a new, empty dbas index}
\usage{
create_dbas_index(escon, aliasName, rfIndex)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{aliasName}{}

\item{rfIndex}{index name for rawfiles index}
}
\description{
Create a new, empty dbas index
}
