% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{create_dbas_index_all}
\alias{create_dbas_index_all}
\title{Create new index for all documents}
\usage{
create_dbas_index_all(escon, rfindex, dateNum = NULL)
}
\arguments{
\item{escon}{ElasticSearch connection object created by `elastic::connect`}

\item{rfindex}{Index name of msrawfiles index}

\item{dateNum}{Provide a specific version code (YYMMDD), default is NULL (current date)}
}
\value{
True (invisibly) when successful
}
\description{
Create new index for all documents
}
