% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{create_index_all}
\alias{create_index_all}
\title{Create new index for DBAS results documents}
\usage{
create_index_all(escon, rfindex, type, dateNum = NULL)
}
\arguments{
\item{escon}{ElasticSearch connection object created by `elastic::connect`}

\item{rfindex}{Index name of msrawfiles index}

\item{type}{type of processing, 'nts' or 'dbas'}

\item{dateNum}{Provide a specific version code (YYMMDD), default is NULL (current date)}
}
\value{
True (invisibly) when successful
}
\description{
Create new index for DBAS results documents
}
