% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_ufid_lib}
\alias{create_ufid_lib}
\title{Create ufid-library}
\usage{
create_ufid_lib(pth)
}
\arguments{
\item{pth}{Path to the where the SQLite file should be saved}
}
\value{
Prints the list of tables and returns full path to library
}
\description{
Produce an empty ufid-library (SQLite database file) at a specified location.
}
