% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{createAllStructures}
\alias{createAllStructures}
\title{Create images of structural formulae from the spectral library}
\usage{
createAllStructures(databasePath, targetDir)
}
\arguments{
\item{databasePath}{Path to spectral library (CSL, SQLite format)}

\item{targetDir}{Path where the PNGs are to be saved}
}
\description{
Uses the spectral library (CSL) to obtain SMILES codes
to create PNG images of the structural formulae for all compounds.
}
