% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-utilities.R
\name{createBackupMsrawfiles}
\alias{createBackupMsrawfiles}
\title{Create a backup of an msrawfiles index}
\usage{
createBackupMsrawfiles(msrawfilesName, overwrite = FALSE)
}
\arguments{
\item{msrawfilesName}{Name of index to back up}

\item{overwrite}{Logical, should an existing index be overwritten (default is false)}
}
\value{
Returns the name of the new table
}
\description{
Will create a backup index with the current date in the name (`<msrawfilesName>_backup_<YYYYMMDD>`). It will not
overwrite an index unless overwrite = TRUE. The backup index is immediately closed so that it does not interfere.
}
