% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{dbas_index_from_alias}
\alias{dbas_index_from_alias}
\title{Get current index name based on the alias}
\usage{
dbas_index_from_alias(aliasName, dateNum = NULL)
}
\arguments{
\item{aliasName}{}

\item{dateNum}{}
}
\description{
The current date will be used in the index name
}
