% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningNewBatches}
\alias{dbaScreeningNewBatches}
\title{Process all measurement files which have not yet been processed by DBAS}
\usage{
dbaScreeningNewBatches(msrawfileIndex, saveDirectory, numParallel = 1)
}
\arguments{
\item{msrawfileIndex}{Name of index where processing information is stored}

\item{saveDirectory}{Location where results should be saved}

\item{numParallel}{For local parallel processing via future}
}
\description{
To determine which files have not yet been processed, the function
collects all directories in `msrawfiles` and looks in the `feature` indices
to see which are missing.
}
