% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningOneBatch}
\alias{dbaScreeningOneBatch}
\title{Process one batch of measurement files by DBAS}
\usage{
dbaScreeningOneBatch(msrawfileRecords, saveDirectory)
}
\arguments{
\item{saveDirectory}{Location where results should be saved}
}
\description{
Perform file scanning (DBAS), convert the result to the NTSPortal `featureRecord` format and save it
to as a JSON.
}
