% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningSelectedBatches}
\alias{dbaScreeningSelectedBatches}
\title{DBAS process selected measurement files via batch directory}
\usage{
dbaScreeningSelectedBatches(
  msrawfileIndex,
  batchDirs,
  saveDirectory,
  numParallel = 1
)
}
\arguments{
\item{msrawfileIndex}{Name of index where processing information is stored}

\item{batchDirs}{Directory of measurement files (recursive, multiple permitted)}

\item{saveDirectory}{Location where SLURM job files should be saved}

\item{numParallel}{For local parallel processing via future}
}
\description{
Choose which batches to process by selecting the directory where the measurement files are stored.
}
