% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningSelectedBatchesSlurm}
\alias{dbaScreeningSelectedBatchesSlurm}
\title{DBAS process batches by directory using the workload manager}
\usage{
dbaScreeningSelectedBatchesSlurm(
  msrawfileIndex,
  batchDirs,
  saveDirectory,
  email
)
}
\arguments{
\item{msrawfileIndex}{Name of index where processing information is stored}

\item{batchDirs}{Directory of measurement files (recursive, multiple permitted)}

\item{saveDirectory}{Location where SLURM job files should be saved}

\item{email}{Address for SLURM notifications}
}
\description{
Choose which batches to process by selecting the directory where the measurement files are stored. 
The function will create necessary files for submission of the job to the workflow manager SLURM.
}
\details{
Three files are created, the records of all files to process, organized into batches (.RDS); the R script which
SLURM needs to run, and the SLURM job file. The command needed to start the process is given as a message (you
may need to switch to a SLURM-enabled server).
}
\examples{
\dontrun{
userEmail <- "example@bafg.de"
dirResults <- "processingResults"
msrawfileIndexName <- "ntsp25.1_msrawfiles"
library(ntsportal)
connectNtsportal()
file.remove(list.files(dirResults, f = T))

dbaScreeningSelectedBatchesSlurm(
  msrawfileIndex = msrawfileIndexName,
  batchDirs = "/root/dir/all/msrawfiles",
  saveDirectory = dirResults,
  email = userEmail
)
}

}
