% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningUnprocessedBatchesSlurm}
\alias{dbaScreeningUnprocessedBatchesSlurm}
\title{Process new batches (DBAS) using SLURM}
\usage{
dbaScreeningUnprocessedBatchesSlurm(msrawfileIndex, saveDirectory, email)
}
\arguments{
\item{msrawfileIndex}{Name of index where processing information is stored}

\item{saveDirectory}{Location where SLURM job files should be saved}

\item{email}{Address for SLURM notifications}
}
\description{
Only new batches are selected, see `dbaScreeningNewBatches()`. The function creates 
the necessary file to submit the processing job to a SLURM workload manager. See `dbaScreeningSelectedBatchesSlurm()`
}
