% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R,
%   R/dbInterface-PythonDbComm.R
\docType{methods}
\name{getTableAsRecords}
\alias{getTableAsRecords}
\alias{ping}
\alias{show,PythonDbComm-method}
\alias{ping,PythonDbComm-method}
\title{Get a table as records}
\usage{
getTableAsRecords(
  dbComm,
  tableName,
  searchBlock = list(),
  recordConstructor = newNtspRecord
)

ping(dbComm)

\S4method{show}{PythonDbComm}(object)

\S4method{ping}{PythonDbComm}(dbComm)
}
\arguments{
\item{recordConstructor}{function to construct new records}

\item{queryBlock}{list coercible to json for elasticSearch REST API (Query DSL)}
}
\value{
A list of ntspRecord objects

TRUE when connection active
}
\description{
Get a table as records

Test the DB-Connection
}
