% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R,
%   R/dbInterface-PythonDbComm.R
\name{deleteTable}
\alias{deleteTable}
\alias{deleteTable,PythonDbComm-method}
\title{Delete a table}
\usage{
deleteTable(dbComm, tableName)

\S4method{deleteTable}{PythonDbComm}(dbComm, tableName)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database}
}
\description{
Delete a table
}
