% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_comp_groups}
\alias{es_add_comp_groups}
\title{Add compound classification data from spectral library to ntsp index}
\usage{
es_add_comp_groups(escon, sdb, index)
}
\arguments{
\item{escon}{}

\item{sdb}{connection to spec lib}

\item{index}{}
}
\value{
True (invisibly)
}
\description{
Add compound classification data from spectral library to ntsp index
}
