% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_identifiers}
\alias{es_add_identifiers}
\title{Add formula, inchi and inchikey to docs in ntsp}
\usage{
es_add_identifiers(escon, sdb, index, filenames = "all")
}
\arguments{
\item{filenames}{Limit the update to results documents from specific files defined by 'filename' field.
If 'all' (default) all dbas documents are updated.}
}
\description{
Add formula, inchi and inchikey to docs in ntsp
}
