% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_add_tag_to_ids}
\alias{es_add_tag_to_ids}
\title{Add a tag to selected docs}
\usage{
es_add_tag_to_ids(escon, index, ids, tagToAdd)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}

\item{ids}{character of document IDs, max. 65536}

\item{tagToAdd}{your comment in the field tag ("yourtexthere")}
}
\value{
Integer. Number of docs updated (0 for no assignment)
}
\description{
Docs are selected by passing a vector of IDs
}
