% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_add_ufid2_to_ids}
\alias{es_add_ufid2_to_ids}
\title{Add ufid2 entry to a list of features in elasticsearch}
\usage{
es_add_ufid2_to_ids(escon, index, ufid2_to_add, ids_for_update)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}

\item{ufid2_to_add}{integer length 1 ufid2 to add}

\item{ids_for_update}{character of document IDs, max. 65536}
}
\value{
TRUE if successful
}
\description{
Add ufid2 entry to a list of features in elasticsearch
}
