% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_add_ufid_to_ids}
\alias{es_add_ufid_to_ids}
\title{Add ufid entry to a list of features in elasticsearch}
\usage{
es_add_ufid_to_ids(
  escon,
  index,
  ufid_to_add,
  ids_for_update,
  ufidType = "ufid",
  verbose = FALSE
)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}

\item{ufid_to_add}{integer length 1 ufid to add}

\item{ids_for_update}{character of document IDs, max. 65536}

\item{ufidType}{either ufid or ufid2}
}
\value{
TRUE if update successful
}
\description{
This function often leads to errors
}
