% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_value}
\alias{es_add_value}
\title{Add a value to any field in ES index}
\usage{
es_add_value(escon, esindex, queryBody, listToAdd = NULL, ...)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{esindex}{Index name or index-pattern}

\item{queryBody}{query in list format}

\item{listToAdd}{List of values to add, (each must be length 1). Field 
must already exist in the index. Add values as a list or as ... arguments (not both)}

\item{...}{Name of field and values to add (each must be length 1). Field 
must already exist in the index. Add values as a list or as ... arguments (not both)}
}
\value{
API response (list)
}
\description{
Will add data to docs. At the moment only works for length 1 
values and will overwrite any current value or values in the field.
}
\details{
This function cannot yet deal with appending data. Values will replace the
current content of the field.
}
