% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucid_assignment.R
\name{es_assign_ucids}
\alias{es_assign_ucids}
\title{Assign ucids to all documents}
\usage{
es_assign_ucids(escon, index = "g2_nts*", rtlim = 0.2)
}
\arguments{
\item{index}{should be all indices together otherwise ucids won't match across indices}

\item{rtlim}{pre-eliminate ufids from consideration which have rt difference higher than this}
}
\value{
TRUE if ran to completion
}
\description{
Totally new assignment, will delete previous ucids
}
\details{
build distance matrix of all ufids against eachother
each element of the matrix says how often the two ufids
are in the same component per batch.
}
