% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{es_check_docs_fields}
\alias{es_check_docs_fields}
\title{Check for consistency of documents in an es index.}
\usage{
es_check_docs_fields(escon, index)
}
\arguments{
\item{escon}{}

\item{index}{}
}
\description{
Check for consistency of documents in an es index.
}
