% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_check_feat}
\alias{es_check_feat}
\title{Check feature}
\usage{
es_check_feat(escon, index, id)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}

\item{id}{character ID length 1}
}
\value{
TRUE if feature ok
}
\description{
Checks that feature has mz.
}
