% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{es_error_handler}
\alias{es_error_handler}
\title{Handle errors from ElasticSearch}
\usage{
es_error_handler(thisCnd)
}
\arguments{
\item{thisCnd}{Error condition thrown in the `tryCatch` context}
}
\value{
No return value
}
\description{
Errors from ElasticSearch are returned by package 'elastic' as text. This
function will take the error condition and take appropriate action.
}
