% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_feat_from_ids}
\alias{es_feat_from_ids}
\title{Get list of features from ntsp with document-IDs}
\usage{
es_feat_from_ids(escon, index, esids)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}

\item{esids}{character vector of IDs}
}
\value{
list of Feature objects
}
\description{
Can one or more IDs
}
