% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_get_comps}
\alias{es_get_comps}
\title{Get list of all compounds in an ntsp index}
\usage{
es_get_comps(escon, index, addQuery = NULL)
}
\arguments{
\item{escon}{}

\item{index}{}

\item{addQuery}{query content in the form of a list.}
}
\description{
Option to add a query to limit the documents
}
\examples{
\dontrun{
termsSearch <- list(
  terms = list(
    filename = as.list(filenames)
  )
)

comps <- es_get_comps(escon, index, termsSearch)
}


}
