% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{es_move_alias}
\alias{es_move_alias}
\title{Change alias of an index to a new index.}
\usage{
es_move_alias(escon, indexName, aliasName, closeAfter = FALSE)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{indexName}{Name of index}

\item{aliasName}{Name of alias}

\item{closeAfter}{logical, should the previous index, to which the alias 
linked to prior to the change, be closed?}
}
\value{
Passes on response from ES (acknowledged field, TRUE or FALSE), invisibly
}
\description{
Will delete the previous alias
}
