% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{es_no_duplicates}
\alias{es_no_duplicates}
\title{Check if there are duplicates in the es database}
\usage{
es_no_duplicates(escon, index)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Elasticsearch index name}
}
\value{
TRUE if no duplicates found.
}
\description{
A cardinality estimation is made. If the number of different docs
(based on mz, rt, intensity, pol, start and station) and the
number of docs is the same, it is assumed no duplicates. The estimation
is done within 5% accuracy since the cardinality computation is fuzzy
(HyperLogLog++ algorithm). The question here is, is it worth it.
}
