% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{es_remove_by_filename}
\alias{es_remove_by_filename}
\title{Removal of documents from dbas documents based on filename}
\usage{
es_remove_by_filename(escon, index, filenames)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{index}{Elasticsearch index name}

\item{filenames}{Vector of filenames to delete}
}
\value{
TRUE if successful
}
\description{
Removal of documents from dbas documents based on filename
}
