% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_remove_ufid}
\alias{es_remove_ufid}
\title{Delete all ufid entries for a ufid}
\usage{
es_remove_ufid(escon, index, ufidToDelete)
}
\arguments{
\item{escon}{}

\item{index}{}

\item{ufidToDelete}{}
}
\description{
Delete all ufid entries for a ufid
}
