% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_res_feat_list}
\alias{es_res_feat_list}
\title{Reformat results from elastic into a list of features}
\usage{
es_res_feat_list(res, fields = c("mz", "pol", "rt", "chrom_method"))
}
\arguments{
\item{fields}{which fields should be included in the features (at leaste mz,
rt, pol and chrom_method must be included)}
}
\value{
list of feature objects
}
\description{
Given a result list from elastic::Search, reformat this into a list of
features of the feature class.
}
