% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{es_search_paged}
\alias{es_search_paged}
\title{Get search results for more than 10000 docs by pagination.}
\usage{
es_search_paged(escon, index, searchBody = NULL, sort, totalSize = Inf, ...)
}
\arguments{
\item{escon}{}

\item{index}{}

\item{searchBody}{search body, if NULL, will return all docs}

\item{sort}{Sort argument passed onto elastic::Search. Defines which field 
the results are sorted by (best if this is unique for all docs to avoid ties) 
currently can only be one field.}

\item{totalSize}{}

\item{...}{further arguments to elastic::Search}
}
\description{
Get search results for more than 10000 docs by pagination.
}
