% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_testing.R
\name{es_test_fpfn}
\alias{es_test_fpfn}
\title{Compute false positives and false negatives in ufid and ucid assignment}
\usage{
es_test_fpfn(escon, udb, index)
}
\arguments{
\item{index}{}
}
\value{
list of metrics used to assess quality of ufid and ucid assignments
}
\description{
Used for testing purposes
}
