% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_ufid_gap_fill}
\alias{es_ufid_gap_fill}
\title{Assign ufids to features without MS2 (gap-filling)}
\usage{
es_ufid_gap_fill(
  escon,
  index,
  ufid_to_fill,
  min_number = 2,
  mztol_gap_fill_mda = 5,
  rttol_gap_fill_min = 0.3
)
}
\arguments{
\item{min_number}{minimum number of MS2 ufids needed in the cluster to assign
ufid to other features in that cluster}

\item{rttol_gap_fill_min}{}
}
\value{
Integer. Number of features assigned (0 for no assignment)
}
\description{
For a given ufid, will look for features of the same batch with similar
mz, rt and eic (chrom. peak shape) and assign these to the same ufid.
}
\details{
Candidate features (based on mz and rt) are joined to already assigned features
and all features are clustered based on dtw correlation of EICs followed by DBSCAN clustering.
}
