% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{esSearchPaged}
\alias{esSearchPaged}
\title{Get search results for more than 10000 docs by pagination.}
\usage{
esSearchPaged(
  indexName,
  searchBody = list(query = list(match_all = stats::setNames(list(), character(0)))),
  sort,
  totalSize = Inf,
  ...
)
}
\arguments{
\item{indexName}{Name of ElasticSearch index}

\item{searchBody}{search body, default will return all docs}

\item{sort}{Sort argument passed onto elastic::Search. Defines which field 
the results are sorted by (best if this is unique for all docs to avoid ties) 
currently can only be one field and may not be '_id'.}

\item{totalSize}{}

\item{...}{further arguments to elastic::Search, asdf argument does not work for now.}
}
\value{
ElasticSearch API response as a list
}
\description{
local function
}
