% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-enrichPolicies.R
\name{executeEnrichPolicy}
\alias{executeEnrichPolicy}
\title{Refresh an existing enrich policy}
\usage{
executeEnrichPolicy(policyName, asBackgroundTask = TRUE)
}
\arguments{
\item{policyName}{Name of policy}

\item{asBackgroundTask}{Boolean, should it run in the background?}
}
\value{
Task ID
}
\description{
Refresh an existing enrich policy
}
