% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{extract_spec_file}
\alias{extract_spec_file}
\title{Extract chromatograms and spectra on a file-wise basis}
\usage{
extract_spec_file(docsOneFile, peakListLong, eicExtractWidth)
}
\arguments{
\item{docsOneFile}{list of docs all from one measurement file}

\item{peakListLong}{peak list in long format (from proco object)}

\item{eicExtractWidth}{extraction width to get EIC (Da)}
}
\value{
list of docs with spectra and chromatograms included (when available)
}
\description{
Takes a list of ntsp docs organised by sample and extracts chromatograms
and spectra from the rawfiles and attaches these to the docs. This is
only for the nts processing side becuase it relies on the peaklist
}
\details{
This function is slow because the file must be opened and a lot of loops need
to be made to extract peaks.
}
