% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{feature_no_ufid}
\alias{feature_no_ufid}
\title{Return one non-aligned feature (without a ufid)}
\usage{
feature_no_ufid(escon, index)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}
}
\value{
returns a ntsportal::feature object
}
\description{
Return one non-aligned feature (without a ufid)
}
