% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{feature_no_ufid_cbz}
\alias{feature_no_ufid_cbz}
\title{Return a feature, annotated with Carbamazepine, without a ufid}
\usage{
feature_no_ufid_cbz(escon, index)
}
\arguments{
\item{escon}{Elasticsearch connection object created by elastic::connect}

\item{index}{Elasticsearch index name}
}
\value{
returns a ntsportal::feature object
}
\description{
Return a feature, annotated with Carbamazepine, without a ufid
}
