% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{find_templateid}
\alias{find_templateid}
\title{Get ID based on search parameters}
\usage{
find_templateid(
  escon,
  rfindex,
  isBlank = FALSE,
  polarity,
  station,
  matrix = "spm"
)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{isBlank}{boolean default is FALSE}

\item{polarity}{}

\item{matrix}{character default is "spm"}
}
\value{
string templateID
}
\description{
Get ID based on search parameters
}
