% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-findTemplate.R
\name{findTemplateId}
\alias{findTemplateId}
\title{Get ID based on search parameters}
\usage{
findTemplateId(
  rfIndex,
  blank = FALSE,
  pol = "pos",
  station = "rhein_ko_l",
  matrix = "spm",
  duration = 365
)
}
\arguments{
\item{rfIndex}{index name for rawfiles index}

\item{pol}{Either "pos" or "neg", default "pos"}

\item{station}{Station ID name}

\item{matrix}{character default is "spm"}

\item{duration}{numeric, default 365}

\item{isBlank}{boolean default is FALSE}
}
\value{
ES-ID of template record in msrawfiles index
}
\description{
Get ID based on search parameters
}
