% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{free_gb}
\alias{free_gb}
\title{Return the number of Gb of free memory on linux server}
\usage{
free_gb()
}
\value{
numeric length 1
}
\description{
Return the number of Gb of free memory on linux server
}
