% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-addRecord.R
\name{get_all_ids}
\alias{get_all_ids}
\title{Get all esids from an index}
\usage{
get_all_ids(escon, rfindex)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}
}
\value{
character vector of IDs
}
\description{
The maximum allowed number of return ids is 10000. This function will
hang if there are more than 10000 docs in the index. In that case it must
be re-written to page over all results.
}
