% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{get_field}
\alias{get_field}
\title{Get a field for a set of documents defined by id}
\usage{
get_field(escon, indexName, esids, fieldName, simplify = T, justone = F)
}
\arguments{
\item{escon}{Elasticsearch connection object}

\item{indexName}{}

\item{esids}{Document IDs in the named index}

\item{fieldName}{field to extract}

\item{simplify}{}

\item{justone}{}
}
\description{
Get a field for a set of documents defined by id
}
