% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{get_field_builder}
\alias{get_field_builder}
\title{Function factory just to avoid typing escon and index every time}
\usage{
get_field_builder(escon, index)
}
\arguments{
\item{esids}{}

\item{fieldName}{}

\item{simplify}{}

\item{justone}{}
}
\description{
Function factory just to avoid typing escon and index every time
}
