% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{get_msrawfiles}
\alias{get_msrawfiles}
\title{Just get the whole msrawfiles index}
\usage{
get_msrawfiles(escon, rfindex)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{msrawfiles index name}
}
\value{
entire msrawfiles as an msrawfiles_docs_list
}
\description{
We just collect the whole index and do any further further manipulation in R
The sorting is done on the path field.
}
