% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_next_ufid}
\alias{get_next_ufid}
\title{Return a new, unique ufid number}
\usage{
get_next_ufid(udb, escon, index)
}
\arguments{
\item{udb}{ufid-library connection object produced with `DBI::dbConnect`}

\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Elasticsearch index name}
}
\value{
length one integer of next ufid
}
\description{
The next ufid number is returned, for creating a new cluster. Will double
check that this number is not already present.
}
