% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{get_rare}
\alias{get_rare}
\title{Get rare compounds from a Report}
\usage{
get_rare(repo, min_freq)
}
\arguments{
\item{repo}{ntsworkflow::Report class object}

\item{min_freq}{Minimum number of detections}
}
\value{
Names of compounds with FEWER than the min_freq of detections
}
\description{
Get the compounds which are found fewer than min_freq
}
