% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recall.R
\name{get_time_series}
\alias{get_time_series}
\title{Get data from elastic}
\usage{
get_time_series(
  escon,
  index = "g2_nts_expn",
  station = "KOMO",
  startRange = c("2021-01-01", "2022-01-01"),
  responseField = "intensity",
  ufidLevel = 2,
  form = "long"
)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Name of index must be one of "g2_nts_expn", "g2_nts_bfg"}

\item{station}{Can be one of "KOMO", "FAN", "WIN", "rhein_ko_l"}

\item{startRange}{Two element character vector with date range e.g. c("2021-01-01", "2022-01-01")}

\item{responseField}{Can be one of "intensity", "intensity_normalized", "area", "area_normalized"}

\item{ufidLevel}{Can be 1 or 2. Level 1: mz+rt+ms2, level 2: mz+rt}

\item{form}{Format of output table, can be one of "long", "wide"}
}
\value{
data.frame
}
\description{
Get data from elastic
}
\examples{
\dontrun{
config_path <- "~/projects/config.yml"
ec <- config::get("elastic_connect", file = config_path)
escon <- elastic::connect(host = "10.140.73.204", user = ec$user, pwd = ec$pwd)
index <- "g2_nts_expn"
get_time_series(
  escon,
  station = "KOMO",
  startRange = c("2021-01-01", "2022-01-01"),
  responseField = "intensity"
)
}

}
