% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recall.R
\name{get_time_series_known}
\alias{get_time_series_known}
\title{Get entire time series for one compound at one station}
\usage{
get_time_series_known(
  escon,
  index = "g2_dbas_bfg",
  compound = "Carbamazepine",
  station = "rhein_ko_l",
  pol = "pos",
  matrix = "water",
  calendarInterval = "day"
)
}
\arguments{
\item{escon}{}

\item{index}{}

\item{compound}{}

\item{station}{}

\item{pol}{}

\item{matrix}{}

\item{calendarInterval}{}
}
\description{
Function should be expanded to allow user to choose date range and get data
from more than one compound, more than one station.
}
