% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{get_unevaluated}
\alias{get_unevaluated}
\title{Get files which have not yet been processed}
\usage{
get_unevaluated(escon, rfindex, evalType = "dbas")
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{evalType}{either dbas or dbas_is}
}
\value{
character vector of ids of documents
}
\description{
A processed file will have a time stamp. This function will return the ids
of all files without a time stamp for the given evaluation type.
}
