% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{get_unproc_batches}
\alias{get_unproc_batches}
\title{Get unprocessed files as batches}
\usage{
get_unproc_batches(allDocs, type)
}
\arguments{
\item{allDocs}{entire msrawfiles index as docsList from elasticSearch}

\item{type}{type of processing, 'nts' or 'dbas'}
}
\value{
list of batches, each containing the docs for processing
}
\description{
Get unprocessed files as batches
}
