% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{getDbComm}
\alias{getDbComm}
\title{Make a DbComm interface object using the constructor given by the ntsportal.dbComm option}
\usage{
getDbComm()
}
\description{
Make a DbComm interface object using the constructor given by the ntsportal.dbComm option
}
