% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R,
%   R/dbInterface-PythonDbComm.R
\name{getNrow}
\alias{getNrow}
\alias{getNrow,PythonDbComm-method}
\title{Get the number of rows in a table}
\usage{
getNrow(dbComm, tableName, searchBlock = list())

\S4method{getNrow}{PythonDbComm}(dbComm, tableName, searchBlock = list())
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}
}
\value{
integer
}
\description{
The number of rows (or doc count) is returned for a given table (wildcards allowed). 
The `searchBlock` argument can be used to limit the results to a search query.
}
