% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{getTableAsRecords}
\alias{getTableAsRecords}
\title{Get a table as a list of ntspRecords}
\usage{
getTableAsRecords(
  dbComm,
  tableName,
  searchBlock = list(),
  fields = "*",
  recordConstructor = newNtspRecord
)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}

\item{fields}{Fields to include in the response, wildcards are permitted, default is all fields.}

\item{recordConstructor}{function to construct new records}
}
\value{
A `list` of `ntspRecord` objects
}
\description{
The table is retrieved as `list` of `ntspRecord`s or subclass thereof. The `recordConstructor` can be used to
make a subclass, e.g., to make `msrawfilesRecord`s use `newMsrawfilesRecord`. The default makes the generic `ntspRecord`.
}
