% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{getTableAsTibble}
\alias{getTableAsTibble}
\title{Get a table as a tibble}
\usage{
getTableAsTibble(dbComm, tableName, searchBlock = list())
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}
}
\value{
A tibble
}
\description{
Get a table as a tibble
}
