% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-getTable.R
\name{getTableByEsql}
\alias{getTableByEsql}
\title{Use ES|QL to retrieve a table from NTSPortal}
\usage{
getTableByEsql(esql)
}
\arguments{
\item{esql}{ES|QL string}
}
\description{
ES|QL can be prepared in the "Discover" page of NTSPortal and
}
\details{
ES|QL does not work for nested fields, fields such as ms2, rtt will be ignored
}
\examples{
\dontrun{
getTableByEsql("FROM ntsp25.1_dbas* | WHERE station == \"mosel_139\" | KEEP mz | LIMIT 10")
}

}
