% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-getTable.R
\name{getTableByQuery}
\alias{getTableByQuery}
\title{Retrieve a table from NTSPortal using Query DSL}
\usage{
getTableByQuery(
  tableName,
  searchBlock = list(),
  fields = "*",
  sortField = "no-sort"
)
}
\arguments{
\item{tableName}{Name of table in database, wildcards are permitted}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}

\item{fields}{Fields to include in the response, wildcards are permitted, default is all fields.}

\item{sortField}{field by which to sort the results, length-one character (optionally preceded by '-') or list, see
https://elasticsearch-dsl.readthedocs.io/en/stable/api.html#elasticsearch_dsl.Search.sort}
}
\value{
A tibble
}
\description{
Works similarly to `getTableAsRecords()` but reformats the `ntspRecords` into a `tibble`.
}
\details{
The records can take a long time to reformat, use the searchBlock and fields arguments to 
refine the search as much as possible thereby reducing the size of the data.
}
\seealso{
\link{`getTableAsRecords`}

\link{`getTableByEsql`}
}
