% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{getUniqueValues}
\alias{getUniqueValues}
\title{Get unique values in a field of a table}
\usage{
getUniqueValues(dbComm, tableName, field, maxLength = 10000)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{field}{table column (field) name see mappings}

\item{maxLength}{maximum number of fields}
}
\value{
a vector
}
\description{
Get unique values in a field of a table
}
