% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntspRecords-dataAccess.R
\name{gf}
\alias{gf}
\title{Extract field from docs list}
\usage{
gf(docsSrc, fieldName, value, justone = F)
}
\arguments{
\item{docsSrc}{docs_source object: list returned by elasticSearch API (only the _source part)}

\item{fieldName}{field to extract (length 1 character)}

\item{value}{FUN.VALUE argument for vapply or "unknown", in which case a list is returned}

\item{justone}{logical return only length one and warn if not all the same}
}
\value{
vector or list of the extracted value
}
\description{
a wrapper around vapply or lapply to get values from docs_source object
}
