% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{glue_json}
\alias{glue_json}
\title{Variation of glue::glue for json input}
\usage{
glue_json(...)
}
\arguments{
\item{...}{}
}
\value{
glue object
}
\description{
Works the same as glue but variables are delimited by '[[]]' and linebreaks
are removed in the output.
}
