% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{index_from_alias}
\alias{index_from_alias}
\title{Get current index name based on the alias}
\usage{
index_from_alias(aliasName, dateNum = NULL)
}
\arguments{
\item{aliasName}{Name of alias used in elasticsearch}

\item{dateNum}{Provide a specific version code (YYMMDD), default is NULL (current date)}
}
\value{
Index name used in ElasticSearch
}
\description{
The current date will be used in the index name
}
