% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-ingest.R
\name{ingest}
\alias{ingest}
\title{Ingest of json.gz files into Elasticsearch}
\usage{
ingest(path)
}
\arguments{
\item{path}{Path to single json.gz file or directory with json.gz files.}
}
\value{
list with one top-level entry per batch (json.gz file). The second level has one entry per
alias and this contains a character vector of the index name associated with that alias.
}
\description{
The processing output is a gzip-compressed json file for each batch (or more than one for large batches). These
are ingested into NTSPortal Elasticsearch.
}
