% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{ingest_all_batches}
\alias{ingest_all_batches}
\title{Ingest a series of ntsportal json files}
\usage{
ingest_all_batches(
  escon,
  rfindex,
  resDir,
  configPath = "~/config.yml",
  ingestScriptPath,
  type,
  pauseTime = 2
)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{msrawfiles index name}

\item{resDir}{Where are the jsons found (must be compressed as json.gz)}

\item{configPath}{Credentials for ntsportal}

\item{ingestScriptPath}{Path where the ingest.sh script is found}

\item{type}{type of processing done, 'nts' or 'dbas'}

\item{pauseTime}{how much time to wait between uploading and checking, make 
sure you give it a lot of pause time for large uploads with many batches}
}
\value{
Successfully uploaded file paths
}
\description{
Ingest a series of ntsportal json files
}
