% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{ingest_ntspl}
\alias{ingest_ntspl}
\title{Ingest an ntsportal json file}
\usage{
ingest_ntspl(escon, ntsplJsonPath, configPath, ingestScriptPath, pauseTime = 2)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{ntsplJsonPath}{Path to json file, the filename must contain the alias into which
the data must be ingested using the delimiters '-inda-' and '-indn-'}

\item{configPath}{Config file where the credentials for signing into 
elasticSearch are found, see ntsportal-wiki}

\item{ingestScriptPath}{Path where the ingest.sh script is found}

\item{pauseTime}{}
}
\value{
Will return true if all data has been ingested successfully, false otherwise
}
\description{
Ingest an ntsportal json file
}
