% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-ingest.R
\name{ingestJson}
\alias{ingestJson}
\title{Ingest of json.gz files into Elasticsearch}
\usage{
ingestJson(json_path)
}
\arguments{
\item{json_path}{Path to single json file or directory with json files.}
}
\value{
Returns pairs for all unique aliases and the respective indeces
}
\description{
Ingest json.gz files into Elasticsearch
}
\details{
Required python packages:
reticulate::py_install('elasticsearch')
reticulate::py_install('pandas')
reticulate::py_install('tqdm')
reticulate::py_install('pyyaml')
}
