% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{isTable}
\alias{isTable}
\title{Does a table exist?}
\usage{
isTable(dbComm, tableName)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database}
}
\value{
logical, true if table exists
}
\description{
Does a table exist?
}
