% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{make_ntspl}
\alias{make_ntspl}
\title{Title}
\usage{
make_ntspl(proco, coresBatch = 1)
}
\arguments{
\item{proco}{}

\item{coresBatch}{number of cores to use for parallelization}
}
\description{
Convert a process output object to a process list object (procl)
}
