% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeMappingsTbl}
\alias{makeMappingsTbl}
\title{Turn mappings into a tibble for documentation}
\usage{
makeMappingsTbl(mappingType)
}
\arguments{
\item{mappingType}{name of mapping type, e.g. "feature"}
}
\value{
tbl_df of mappings with metadata
}
\description{
Turn mappings into a tibble for documentation
}
