% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-versionIncrementer.R
\name{msrawfilesSetVersion}
\alias{msrawfilesSetVersion}
\title{Change msrawfiles to a new version}
\usage{
msrawfilesSetVersion(msrawfilesName, version)
}
\arguments{
\item{msrawfilesName}{Name of `msrawfiles` table}

\item{version}{New version number}
}
\description{
To move to a new version of NTSPortal, `msrawfiles` needs to be updated with
new alias names.
}
\examples{
\dontrun{
msrawfilesSetVersion("ntsp25.1_msrawfiles", "25.2")
}
}
