% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_class.R
\name{new_feature}
\alias{new_feature}
\title{Create feature}
\usage{
new_feature(
  pol = character(),
  mz = double(),
  rt = double(),
  es_id = character(),
  chrom_method = character(),
  ...
)
}
\arguments{
\item{pol}{polarity of the feature}

\item{mz}{mass of the feature}

\item{rt}{retentiontime of the feature}

\item{chrom_method}{which method was used to measure}

\item{...}{}
}
\description{
Create feature
}
