% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_assignment_class.R
\name{new_ufid_assignment}
\alias{new_ufid_assignment}
\title{New ufid}
\usage{
new_ufid_assignment(x = integer(), level = integer())
}
\arguments{
\item{x}{ID for ufid}

\item{level}{ufid level}
}
\value{
ufid_assignment object
}
\description{
New ufid
}
