% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionToElasticsearch.R
\name{newPythonComm}
\alias{newPythonComm}
\title{Create an Elasticsearch Python connection}
\usage{
newPythonComm()
}
\value{
class pythonComm
}
\description{
Create an Elasticsearch Python connection
}
