% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{norm_ms1}
\alias{norm_ms1}
\title{Normalize ms1 spectrum to the maximum intensity}
\usage{
norm_ms1(x, precursorMz, noiselevel = 0.01)
}
\arguments{
\item{x}{Spectrum as a data.frame with columns mz and int}

\item{noiselevel}{Intensity level to set as noise}

\item{precursormz}{Molecular ion precursor m/z}

\item{precursorInt}{Molecular ion precursor intensity}
}
\value{
Normalized spectrum as a data.frame
}
\description{
Normalize ms1 spectrum to the maximum intensity
}
