% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{norm_ms2}
\alias{norm_ms2}
\title{Normalize ms2 spectrum to the maximum intensity}
\usage{
norm_ms2(x, precursormz, mztol = 0.02, noiselevel = 0.1)
}
\arguments{
\item{x}{Spectrum as a data.frame with columns mz and int}

\item{precursormz}{Molecular ion precursor m/z}

\item{mztol}{Tolerance of m/z}

\item{noiselevel}{Intensity level to set as noise}
}
\value{
Normalized spectrum as a data.frame
}
\description{
Normalize ms2 spectrum to the maximum intensity
}
