% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{norm_rf_paths}
\alias{norm_rf_paths}
\title{Normalize all paths in the msrawfiles index}
\usage{
norm_rf_paths(escon, rfindex)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}
}
\value{
Returns TRUE when completed (invisibly)
}
\description{
Normalize all paths in the msrawfiles index
}
