% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{ntScreeningOneBatch}
\alias{ntScreeningOneBatch}
\title{Process one batch of measurement files by NTS}
\usage{
ntScreeningOneBatch(msrawfileRecords, saveDirectory)
}
\arguments{
\item{saveDirectory}{Location where results should be saved}
}
\description{
Perform file scanning (NTS), convert the result to the NTSPortal `featureRecord` format and save it
to RDS.
}
