% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntsportal-package.R
\docType{package}
\name{ntsportal-package}
\alias{ntsportal-package}
\title{ntsportal: ntsportal - a database for non-target-screening of surface water}
\description{
Provides functionality for building and managing the NTSPortal database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ntsportal.bafg.de}
}

}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Kevin S. Jewell \email{jewell@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Ole Lessmann \email{lessmann@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
}

Other contributors:
\itemize{
  \item Jonas Skottnik \email{skottnik@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [contributor]
  \item Franziska Prodöhl \email{prodoehl@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [contributor]
  \item Nina Hermes \email{hermes@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [contributor]
  \item Arne Wick \email{wick@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [reviewer]
}

}
\keyword{internal}
