% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R,
%   R/dbInterface-PythonDbComm.R
\docType{methods}
\name{ping}
\alias{ping}
\alias{ping,PythonDbComm-method}
\title{Test the DB-Connection}
\usage{
ping(dbComm)

\S4method{ping}{PythonDbComm}(dbComm)
}
\arguments{
\item{dbComm}{DbComm connection object}
}
\value{
TRUE when connection active
}
\description{
Test the DB-Connection
}
