% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_ms2}
\alias{plot_ms2}
\title{Plot an ms2 spectrum for viewing}
\usage{
plot_ms2(ms2Spec)
}
\arguments{
\item{ms2Spec}{data.frame with columns mz and int}
}
\description{
Plot an ms2 spectrum for viewing
}
