% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{proc_batch_nts}
\alias{proc_batch_nts}
\title{Process a batch of files using ntsworkflow}
\usage{
proc_batch_nts(
  escon,
  rfindex,
  esids,
  tempsavedir,
  ingestpth,
  configfile,
  coresBatch,
  noIngest = FALSE
)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{Elasticsearch index name for msrawfiles index}

\item{esids}{ElasticSearch document IDs (character)}

\item{tempsavedir}{}

\item{ingestpth}{}

\item{configfile}{}

\item{coresBatch}{}

\item{noIngest}{}
}
\description{
Process a batch of files using ntsworkflow
}
