% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{proc_batches_nts}
\alias{proc_batches_nts}
\title{Process batches and save results all steps}
\usage{
proc_batches_nts(batches, saveDir, coresTotal = 1, saveIntermed = F)
}
\arguments{
\item{batches}{ntsp msrawfiles as an msrawfiles_docs_list, as provided by elasticsearch API}

\item{saveDir}{where to save files}

\item{coresTotal}{number of cores to use}

\item{saveIntermed}{save intermediate results for bugfixing}
}
\value{
vector of paths to newly created json files
}
\description{
This function combines all the functions starting from the parameters saved 
in an msrawfiles_docs_list (from msrawfiles index) until the
}
