% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{proc_doc_pp}
\alias{proc_doc_pp}
\title{Process a single file using a list of settings}
\usage{
proc_doc_pp(dsrc)
}
\arguments{
\item{dsrc}{document source ntsp_docs_source}
}
\value{
a list with 3 elements, named pl (peaklist), rf (rawfile), 
and set (settings)
}
\description{
The settings used is a document source of the msrawfiles index
}
