% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{proc_esid}
\alias{proc_esid}
\title{Process an MS measurement file}
\usage{
proc_esid(escon, rfindex, esid, compsProcess = NULL)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{esid}{ElasticSearch document IDs in the rawfiles index (character)}
}
\value{
an object of class ntsworkflow::Report
}
\description{
Takes an esid from the rawfiles index and carries out dbas processing (see details) for 
that file, returns ntsworkflow::Report class object with results.
}
\details{
# dbas processing
dbas processing means to take mzXML measurement file and analyse for compounds found
in the spectral library, peaks are integrated and spectra are extracted and saved
in the Report object.
}
