% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{proc_nts_new}
\alias{proc_nts_new}
\title{Process batches with new files}
\usage{
proc_nts_new(msrawfiles, saveDir, coresTotal = 1)
}
\arguments{
\item{msrawfiles}{}

\item{saveDir}{}

\item{coresBatch}{}
}
\value{
vector of paths to newly created json files
}
\description{
Process batches with new files
}
