% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_mappings.R
\name{put_nts_index}
\alias{put_nts_index}
\title{Create an nts type index}
\usage{
put_nts_index(escon, index)
}
\arguments{
\item{escon}{elasticsearch connection object created with elastic::connect}

\item{index}{Name of the index you wish to create}
}
\description{
The index mappings are saved under inst/extdata as json files.
}
