% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_mappings.R
\name{put_ntsp_index}
\alias{put_ntsp_index}
\title{Create an nts type index}
\usage{
put_ntsp_index(escon, mappingType = "nts", indexName)
}
\arguments{
\item{escon}{elasticsearch connection object created with elastic::connect}

\item{mappingType}{type of mapping (nts only at the moment)}

\item{indexName}{Name of the index you wish to create}
}
\description{
The index mappings are saved under inst/extdata as json files.
}
