% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_mappings.R
\name{put_spectral_library_index}
\alias{put_spectral_library_index}
\title{Send index creation request for spectral library index}
\usage{
put_spectral_library_index(escon, index)
}
\arguments{
\item{escon}{}

\item{index}{}
}
\description{
Send index creation request for spectral library index
}
