% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-PythonDbComm.R
\docType{class}
\name{PythonDbComm-class}
\alias{PythonDbComm-class}
\title{A DbComm interface using the python back-end}
\description{
A DbComm interface using the python back-end
}
\section{Slots}{

\describe{
\item{\code{client}}{A python elasticsearch client object}

\item{\code{dsl}}{The python elasticsearch.dsl import module}
}}

