% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-PythonDbComm.R
\name{PythonDbComm}
\alias{PythonDbComm}
\title{Build Python DB communications interface}
\usage{
PythonDbComm(ring = "ntsportal")
}
\arguments{
\item{ring}{The name of the keyring where NTSPortal credentials are stored, see `connectNtsportal()`}
}
\value{
PythonDbComm
}
\description{
A generator for the `PythonDbComm` class.
}
