% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-addRecord.R
\name{rawfiles_missing}
\alias{rawfiles_missing}
\title{Check if any files are missing}
\usage{
rawfiles_missing(escon, rfindex)
}
\arguments{
\item{escon}{ElasticSearch connection object created by `elastic::connect`}

\item{rfindex}{index name for rawfiles index}
}
\value{
Return paths of files that do not exist
}
\description{
Check if any files are missing
}
