% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{record_proc}
\alias{record_proc}
\title{Record end of processing}
\usage{
record_proc(escon, rfindex, pathRawfiles, type)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{msrawfiles index name}

\item{pathRawfiles}{The paths of the rawfiles (used as primary key to link msrawfiles with results)}

\item{type}{type of processing done, 'nts' or 'dbas'}
}
\value{
TRUE if successful
}
\description{
Record end of processing
}
