% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{remove_na_doc}
\alias{remove_na_doc}
\title{Remove fields in a doc that are or contain NAs}
\usage{
remove_na_doc(doc)
}
\arguments{
\item{doc}{A single document (of a NT feature) in the form of a list}
}
\value{
list document with NAs removed
}
\description{
Remove fields in a doc that are or contain NAs
}
