% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{removeFieldFromTable}
\alias{removeFieldFromTable}
\title{Remove a field from a Table}
\usage{
removeFieldFromTable(dbComm, tableName, field, searchBlock = list())
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{field}{table column (field) name see mappings}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}
}
\description{
Remove a field from a Table
}
