% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{removeRedundantPngs}
\alias{removeRedundantPngs}
\title{Delete redundant PNGs}
\usage{
removeRedundantPngs(databasePath, targetDir)
}
\arguments{
\item{databasePath}{Path to spectral library (sqlite)}

\item{targetDir}{Path where the PNGs are stored}
}
\description{
Delete redundant PNGs
}
