% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{removeValueFromArray}
\alias{removeValueFromArray}
\title{Remove a value from an array field}
\usage{
removeValueFromArray(dbComm, tableName, field, value, searchBlock = list())
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{field}{table column (field) name see mappings}

\item{value}{value to set or add}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}
}
\description{
Remove a value from an array field
}
\seealso{
\link{`addValueToArray`}
}
