% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reset_alignment}
\alias{reset_alignment}
\title{Remove all ufid from NTSPortal-DB and a ufid library}
\usage{
reset_alignment(escon, es_index, ufidLibPath)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{es_index}{index name or pattern}

\item{ufidLibPath}{path to ufid-lib}
}
\value{
TRUE if completed
}
\description{
Remove all ufid from NTSPortal-DB and a ufid library
}
