% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{reset_eval}
\alias{reset_eval}
\title{Reset files in msrawfiles, so that they will be processed again.}
\usage{
reset_eval(escon, rfindex, queryBody, indexType = "dbas", confirm = TRUE)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{queryBody}{The Query DSL code to use to select the docs in msrawfiles which need to be reset (as a `list`).}

\item{indexType}{Elasticsearch index, indicating the type of data stored in the index, can either be "dbas" or "dbas_is"}

\item{confirm}{logical, if True, will ask user to confirm the reset}
}
\description{
If there are errors it may be necessary to reprocess files. This function
will remove the fields "dbas_last_eval" and "dbas_spectral_library_sha256"
so that during the next processing, the files will be processed again.
}
