% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{save_ntspl}
\alias{save_ntspl}
\title{Save}
\usage{
save_ntspl(procList, saveDir, maxSizeGb)
}
\arguments{
\item{procList}{}

\item{saveDir}{}

\item{maxSizeGb}{maximum size of ntspl object before it is split into multiple
parts.}
}
\value{
will save json and return path
}
\description{
Save
}
