% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-dbasFileScanning.R,
%   R/screening-ntsFileScanning.R
\name{scanBatch}
\alias{scanBatch}
\alias{scanBatch.dbasMsrawfilesBatch}
\alias{scanBatch.ntsMsrawfilesBatch}
\title{Extract features from measurement files with an algorithm}
\usage{
scanBatch(msrawfilesBatch, ...)

\method{scanBatch}{dbasMsrawfilesBatch}(msrawfilesBatch, compsToProcess = NULL, showProgress = FALSE)

\method{scanBatch}{ntsMsrawfilesBatch}(msrawfilesRecords, forNtsworkflow = FALSE)
}
\arguments{
\item{msrawfilesBatch}{object inheriting `msrawfilesBatch` (`list` of `msrawfilesRecord`s)}

\item{compsToProcess}{Character vector of compounds names to include in the scanning (default is all compounds in the spectral library)}

\item{showProgress}{Logical, show progress bar? (for interacte use, default false)}

\item{forNtsworkflow}{logical, should the results be prepared for the ntsworkflow non-target app?}
}
\value{
a `dbasResult` (`list` of 7 tables) including `peakList`, `reintegrationResults` etc.

an `ntsResult` object with the tables `peakList`, `sampleList`, `alignmentTable`, `annotationTable`
}
\description{
The spectral library is used to scan mzXML measurment files for compounds using the DBAS algorithm.
The parameters for DBAS and file locations (measurement files and spectral library) are stored the `dbasMsrawfilesBatch`
object.

The  mzXML measurment files are scanned for all features using the NTS algorithm. The spectral library
is used to annotated known features. The parameters for NTS and file locations (measurement files and 
spectral library) are stored in are stored the `ntsMsrawfilesBatch` object.
}
\examples{
\dontrun{
dbComm <- getDbComm()
recs <- getTableAsRecords(
  dbComm, 
  "ntsp25.2_msrawfiles", 
  searchBlock = list(query = list(regexp = list(filename = "Des_19_.._pos.mzXML"))),
  newDbasMsrawfilesRecord
)
recsBlanks <- getTableAsRecords(
  dbComm, 
  "ntsp25.2_msrawfiles", 
  searchBlock = list(query = list(regexp = list(path = ".*mud_pos/BW.*"))), 
  newDbasMsrawfilesRecord
)
newBatch <- newDbasMsrawfilesBatch(c(recs, recsBlanks))
res <- scanBatch(newBatch, "Methyltriphenylphosphonium")
}
}
