% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-fileScanning.R
\name{scanBatchDbas}
\alias{scanBatchDbas}
\title{Process a list of msrawfileRecords}
\usage{
scanBatchDbas(records, compsToProcess = NULL)
}
\arguments{
\item{records}{list of msrawfileRecords}

\item{compsToProcess}{Character vector of compounds names (Default is all compounds in the spectral library)}
}
\value{
a dbasResult object (list of 7 tables) including peakList, reintegrationResults etc.
}
\description{
Process a list of msrawfileRecords
}
\examples{
\dontrun{
dbComm <- getDbComm()
recs <- getTableAsRecords(
  dbComm, 
  "ntsp25.1_msrawfiles", 
  searchBlock = list(query = list(regexp = list(filename = "Des_19_.._pos.mzXML"))),
  newMsrawfilesRecord
)
recsBlanks <- getTableAsRecords(
  dbComm, 
  "ntsp25.1_msrawfiles", 
  searchBlock = list(query = list(regexp = list(path = ".*mud_pos/BW.*"))), 
  newMsrawfilesRecord
)
res <- scanBatchDbas(c(recs, recsBlanks), "Methyltriphenylphosphonium")
}
}
