% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-ntsFileScanning.R
\name{scanBatchNts}
\alias{scanBatchNts}
\title{Scan files by non-target-screening processing}
\usage{
scanBatchNts(msrawfilesRecords, forNtsworkflow = FALSE)
}
\arguments{
\item{msrawfilesRecords}{Files to process as a `list` of `msrawfileRecord`s}

\item{forNtsworkflow}{logical, should the results be prepared for the ntsworkflow non-target app?}
}
\value{
an `ntsResult` object with the tables `peakList`, `sampleList`, `alignmentTable`, `annotationTable`
}
\description{
Scan files by non-target-screening processing
}
