% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{screeningOneBatch}
\alias{screeningOneBatch}
\title{Process one batch of measurement files}
\usage{
screeningOneBatch(msrawfilesBatch, saveDirectory)
}
\arguments{
\item{msrawfilesBatch}{object inheriting `msrawfilesBatch`}
}
\description{
Perform file scanning, converts the result to the NTSPortal `featureRecord` format and saves it
to RDS.
}
