% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{screeningSelectedBatches}
\alias{screeningSelectedBatches}
\title{Process measurement files in selected batch directories}
\usage{
screeningSelectedBatches(
  msrawfileIndex,
  batchDirs,
  saveDirectory,
  numParallel = 1,
  screeningType = "dbas"
)
}
\arguments{
\item{msrawfileIndex}{Name of index where processing information is stored}

\item{batchDirs}{Directory of measurement files (recursive, multiple permitted)}

\item{saveDirectory}{Location where results should be saved}

\item{numParallel}{For local parallel processing via future}

\item{screeningType}{either 'dbas' or 'nts' for library screening and non-target-screening, respectively}
}
\description{
Choose which batches to process by selecting the directory where the measurement files are stored.
}
