% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionToElasticsearch.R
\name{setCredNonInteractive}
\alias{setCredNonInteractive}
\title{Set credentials for accessing NTSPortal ElasticSearch}
\usage{
setCredNonInteractive(ring = "ntsportal", usr = character(), pwd = character())
}
\arguments{
\item{ring}{Keyring name, do not change the default}

\item{usr}{username, enter manually for non-interactive use}

\item{pwd}{password, enter manually for non-interactive use}
}
\description{
For progammatically setting the credentials.
}
