% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubd_distance.R
\name{ubd_new_ufid}
\alias{ubd_new_ufid}
\title{Add new ufid entry by dbscan clustering}
\usage{
ubd_new_ufid(ubd, escon, index, polarity, minPoints1 = 5, minPoints2 = 2)
}
\arguments{
\item{polarity}{}
}
\value{
ufid which was added successful, otherwise 0
}
\description{
Add new ufid entry by dbscan clustering
}
