% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_db_comm.R
\name{udb_clean_spectra}
\alias{udb_clean_spectra}
\title{Clean spectra within ufid db}
\usage{
udb_clean_spectra(udb, ufid, msLevel)
}
\arguments{
\item{msLevel}{}
}
\value{

}
\description{
Clean spectra within ufid db
}
