% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-new-ufid.R
\name{udb_new_ufid}
\alias{udb_new_ufid}
\title{Cluster features and add new ufid entries by dbscan clustering}
\usage{
udb_new_ufid(udb, escon, index, polarity, mzPosition = 0, rtPosition = 0)
}
\arguments{
\item{udb}{ufid database connection object (DBI::dbConnect)}

\item{escon}{NTSPortal connection object (elastic::connect)}

\item{index}{Index name}

\item{polarity}{polarity of features to process}

\item{mzPosition}{minimum mz (Da) for returning new features from NTSPortal}

\item{rtPosition}{minimum RT (min) for returning new features from NTSPortal}
}
\value{
TRUE if the function ended successfully
}
\description{
This is the main function for the clustering stage of NTSP alignment.
New features, which do not yet have a ufid entry, are clustered based on
their similarity (within tolerances, set in the configuration file) in the
m/z, tR and MS2 domains. New ufids are added both to NTSP and ufid-library.
}
