% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_db_comm.R
\name{udb_update}
\alias{udb_update}
\title{Update ufid db based on average aggregated results from ntsportal}
\usage{
udb_update(udb, escon, index, ufid_to_update)
}
\arguments{
\item{udb}{ufid-library connection object produced with `DBI::dbConnect`}

\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Elasticsearch index name}

\item{ufid}{Universal feature ID}
}
\value{
TRUE if successful
}
\description{
The ufid database is updated at the end of a session to
improve it based on statistical averaging of previous data.
Caution is needed here that incorrect data does not contaminate the "gold
standard"
Index to be used should be index-pattern to include all available indexes.
}
