% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_db_comm.R
\name{udb_update_all}
\alias{udb_update_all}
\title{Update the entire ufid library}
\usage{
udb_update_all(udb, escon, index)
}
\arguments{
\item{udb}{ufid-library connection object produced with DBI::dbConnect}

\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Elasticsearch index name}
}
\value{
TRUE if successful
}
\description{
Update the entire ufid library
}
