% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid2_alignment.R
\name{ufid2_alignment}
\alias{ufid2_alignment}
\title{Alignment by m/z and RT}
\usage{
ufid2_alignment(
  escon,
  index = "g2_nts*",
  rtTol = 0.3,
  mzTolmDa = 5,
  minPoints = 5,
  mzRange = c(0, 1e+06),
  rtRange = c(0, 100),
  numCores = 10
)
}
\arguments{
\item{index}{default is "g2_nts*"}

\item{rtTol}{retentiontime tolerance in minutes - default is 0.3min}

\item{mzTolmDa}{mass tolerance in mDa - default is 5mDa}

\item{minPoints}{minimum number of points to build one cluster}

\item{mzRange}{For testing purposes, restrict clustering by m/z, numeric length 2 lower and upper limit, default = c(0, 1e6)}

\item{rtRange}{For testing purposes, restrict clustering by rt, numeric length 2 lower and upper limit, default = c(0, 100)}

\item{numCores}{Number of cores for parallelization, default 10}
}
\value{
TRUE
}
\description{
Function will assign ufid2 to features
}
