% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-enrichPolicies.R
\name{updateEnrichPolicies}
\alias{updateEnrichPolicies}
\title{Update enrich policies in Elasticsearch}
\usage{
updateEnrichPolicies()
}
\value{
a vector of task IDs for all enrich policy execution tasks
}
\description{
Refresh (overwrite and re-execute) the enrich policies as stored in the ntsportal package (inst/enrichPolicies)
}
