% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-linearRegression.R
\name{updateLinearRegressionTable}
\alias{updateLinearRegressionTable}
\title{Update the "analysis" table for spm dashboard}
\usage{
updateLinearRegressionTable(sourceTableName)
}
\arguments{
\item{sourceTableName}{Table in ntsportal which should be analyzed (only dbas_upb)}
}
\description{
The analysis table shows slope of the linear regression model for each compound at each station for yearly composite samples
This can is only computed for yearly composite samples of SPM in the upb dbas index
}
