% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpectralLibrary.R
\name{updateSpectralLibrary}
\alias{updateSpectralLibrary}
\title{Update the CSL spectral library on ntsportal}
\usage{
updateSpectralLibrary(rfTableName, specLibTableName)
}
\arguments{
\item{rfTableName}{Name of the msrawfiles table}

\item{specLibTableName}{Name for the new spectral library table (old table will be overwritten)}
}
\description{
Update the CSL spectral library on ntsportal
}
