% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_class.R
\name{validate_feature}
\alias{validate_feature}
\title{Check feature}
\usage{
validate_feature(x)
}
\arguments{
\item{x}{Feature object}
}
\value{
Unchanged feature object if checks have passed
}
\description{
Check feature
}
\details{
This function will through an error if checks fail.
}
