% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{add_rawfiles}
\alias{add_rawfiles}
\title{Add new MS measurement files to msrawfiles index}
\usage{
add_rawfiles(
  escon,
  rfindex,
  templateId,
  newPaths,
  newStart = "filename",
  newStation = "same_as_template",
  rootMeasDir = "/srv/cifs-mounts/g2/G/G2/HRMS/Messdaten/"
)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{templateId}{Document ID for a document to use as a template}

\item{newPaths}{Character vector of full paths to new rawfiles which are to 
be added, must be mzXML files.}

\item{newStart}{Either "filename" (default), meaning extract the start time 
from the filename or provide a start date as an 8 digit number "YYYYMMDD"}

\item{newStation}{Either be "same_as_template" (default), which will
copy the value from the template document, or "filename", meaning extract
the station and loc values from the code in the filename (will compare to 
other docs in msrawfiles index) or provide a list of length 3 with station, river  
and loc values. See details.}

\item{rootMeasDir}{Root directory where measurement files are located. The function
will look for original (vendor) files in this directory or below. Must end with "/".}
}
\description{
Add new MS measurement files to msrawfiles index
}
\details{
newStation can be either "same_as_template" or "filename" or
a new fixed station name and location (list with fields "station", "river" and "loc")
"loc" being a geopoint with fields "lat" and "lon"
if it is in the filename it will use the dbas_station_regex field in
the index to get the station information from another doc in the index. Station and
river names all lowercase and no spaces and no special characters. 
For the station name use the convention <river_town_position> where position is l, r or m for 
left, right, middle. If there is no obvious town but km
is known, then use the convention <river_km> for the station name. If neither 
town or km is known, use the convention <river_description> where description
is some indication of the location.
}
