% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{change_msrawfile_filename}
\alias{change_msrawfile_filename}
\title{Change filename in the msrawfiles-db and on the filesystem}
\usage{
change_msrawfile_filename(escon, rfindex, oldPath, newPath)
}
\arguments{
\item{escon}{ElasticSearch connection object created by `elastic::connect`}

\item{rfindex}{index name for msrawfiles index}

\item{oldPath}{original path}

\item{newPath}{updated path}
}
\value{
True (invisibly) if successful
}
\description{
This function will change both filename and path of the doc.
}
