% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{change_msrawfile_path}
\alias{change_msrawfile_path}
\title{Change the path of a document in a rawfiles db}
\usage{
change_msrawfile_path(escon, rfindex, oldPath, newPath)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{oldPath}{Current path in rawfiles index, file must exist}

\item{newPath}{New path, file must exist}
}
\value{
TRUE if change was successful (invisibly)
}
\description{
Change the path of a document in a rawfiles db
}
\details{
Both files must exist when making the change. The function takes some time 
because it compares md5-checksums of the two files.
}
