% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{change_msrawfile_path}
\alias{change_msrawfile_path}
\title{Change the path of one document in msrawfiles}
\usage{
change_msrawfile_path(escon, rfindex, oldPath, newPath, checkType = "md5")
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{oldPath}{Current path in rawfiles index, file must exist}

\item{newPath}{New path, file must exist}

\item{checkType}{Method to check that files are the same, either "md5" (default) or "filesize"}
}
\value{
TRUE if change was successful (invisibly)
}
\description{
This is a secure method of changing the path of a doc in msrawfiles. 
Filename is not changed and must remain the same.
}
\details{
Both files must exist when making the change. The function takes some time 
if it compares md5-checksums of the two files. For large numbers of files use the
filesize check.
}
