% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-modification.R
\name{changeMsrawfileFilename}
\alias{changeMsrawfileFilename}
\title{Change filename in the msrawfiles-db and on the filesystem}
\usage{
changeMsrawfileFilename(rfIndex, oldPath, newPath)
}
\arguments{
\item{rfIndex}{index name for msrawfiles table}

\item{oldPath}{Current path in msrawfiles table}

\item{newPath}{New path}
}
\description{
Used to change filename of a measurement file and simultaneously
change the `path` field in `msrawfiles`.
}
