% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-modification.R
\name{changeMsrawfilePath}
\alias{changeMsrawfilePath}
\title{Change the path of one document in msrawfiles}
\usage{
changeMsrawfilePath(rfIndex, oldPath, newPath, checkType = "md5")
}
\arguments{
\item{rfIndex}{index name for rawfiles index}

\item{oldPath}{Current path in rawfiles index, file must exist (1 file)}

\item{newPath}{New path, file must exist (1 file)}

\item{checkType}{Method to check that files are the same, either "md5" (default) or "filesize"}
}
\value{
TRUE if change was successful (invisibly)
}
\description{
This is a secure method of changing the path of a doc in msrawfiles. 
Filename is not changed and must remain the same.
}
\details{
Both files must exist when making the change. The function takes some time 
if it compares md5-checksums of the two files. For large numbers of files use the
filesize check.
}
\examples{
\dontrun{
rfIndex <- "ntsp_index_msrawfiles_unit_tests"
res <- elastic::Search(escon, rfIndex, source = "path")
opths <- sapply(res$hits$hits, function(x) x[["_source"]]$path)
fn <- basename(opths)
npths <- list.files("/scratch/nts/ntsportal_unit_tests/meas_files", f = T)
npths <- nps[sapply(fn, grep, x = opths)]
purrr::walk2(opths, npths, change_msrawfile_path, escon = escon, rfIndex = rfIndex)
}
}
