% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{check_field}
\alias{check_field}
\title{Check that a field exists in all documents in rawfiles index}
\usage{
check_field(escon, rfindex, fieldName, onlyNonBlank = FALSE)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{fieldName}{}

\item{onlyNonBlank}{logical, default FALSE, if TRUE, blanks will be ignored}
}
\value{
TRUE if all docs have the field
}
\description{
Check that a field exists in all documents in rawfiles index
}
