% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles_quality_assurance.R
\name{check_integrity_msrawfiles}
\alias{check_integrity_msrawfiles}
\title{Check consistency of msrawfiles DB}
\usage{
check_integrity_msrawfiles(escon, rfindex, locationRf)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{Elasticsearch index name for msrawfiles index}

\item{locationRf}{Root directory for all rawfiles}
}
\value{
TRUE if successful (invisibly)
}
\description{
Must be done before any processing operation
}
