% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles_quality_assurance.R
\name{check_integrity_msrawfiles_nts}
\alias{check_integrity_msrawfiles_nts}
\title{Check consistency of msrawfiles for nts data}
\usage{
check_integrity_msrawfiles_nts(msrawfilesDocsList)
}
\arguments{
\item{msrawfilesDocsList}{content of entire msrawfiles as a docslist list as returned by elasticsearch API}
}
\description{
Check consistency of msrawfiles for nts data
}
