% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{check_uniformity}
\alias{check_uniformity}
\title{Check that fields are the same for a set of documents}
\usage{
check_uniformity(escon, rfindex, esids, fieldName, onlyNonBlanks = F)
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{esids}{}

\item{fieldName}{}

\item{onlyNonBlanks}{}
}
\description{
Check that fields are the same for a set of documents
}
