% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningProcessBatchScanDbas.R
\name{check_uniformity}
\alias{check_uniformity}
\title{Check that fields are the same for a set of documents}
\usage{
check_uniformity(escon, rfindex, esids, fieldName, onlyNonBlanks = F)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{esids}{ElasticSearch document IDs (character)}

\item{fieldName}{}

\item{onlyNonBlanks}{}
}
\value{
TRUE if the field is uniform accross the docs, false otherwise
}
\description{
Check that fields are the same for a set of documents
}
