% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-Record.R
\name{checkMsrawfiles}
\alias{checkMsrawfiles}
\title{Check an msrawfile index for validity}
\usage{
checkMsrawfiles(indexName = "ntsp_msrawfiles")
}
\arguments{
\item{indexName}{default: ntsp_msrawfiles}
}
\value{
TRUE
}
\description{
This function connects to elasticsearch and checks the given msrawfiles 
index according to the current validation routine.
}
