% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntspRecord-validation.R
\name{checkMsrawfiles}
\alias{checkMsrawfiles}
\title{Check `msrawfiles` for validity}
\usage{
checkMsrawfiles(indexName = "ntsp25.2_msrawfiles")
}
\arguments{
\item{indexName}{default: ntsp_msrawfiles}
}
\value{
TRUE if checks passed
}
\description{
Checks the `msrawfiles` table according to the current validation routine.
}
