% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{checkPngAvailability}
\alias{checkPngAvailability}
\title{Check that PNGs for structural formulae are consistent with the spectral library}
\usage{
checkPngAvailability(databasePath, targetDir)
}
\arguments{
\item{databasePath}{Path to spectral library (sqlite)}

\item{targetDir}{Path where the PNGs are stored}
}
\description{
Check that PNGs for structural formulae are consistent with the spectral library
}
