% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionToElasticsearch.R
\name{connectNtsportal}
\alias{connectNtsportal}
\title{Set ElasticSearch credentials and create ElasticSearch connection object}
\usage{
connectNtsportal()
}
\description{
NTSPortal credentials are used to build the DbComm interface to the database. 
If no credentials are available a pop-up will ask for
user input (in interactive use). Uses the `keyring` package to 
store and retrieve credentials. The default `keyring` backend is used.
When working on Linux in the console, this is the 'file'
backend, on windows, 'wincred'.
The backend can be set manually with `Sys.setenv("R_KEYRING_BACKEND" = "file")`.
}
\details{
Will additionally create the `escon` connection object in the 
global environment for use by all functions (deprecated, use DbComm interface)
}
