% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionToElasticsearch.R
\name{connectNtsportal}
\alias{connectNtsportal}
\title{Create ElasticSearch connection object}
\usage{
connectNtsportal()
}
\description{
If credentials are available, will create the "escon" connection object in the 
global environment for use by all functions. Otherwise will ask for input of
credentials (in interactive use).
}
