% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-convertToRecord.R
\name{convertToRecord.dbasResult}
\alias{convertToRecord.dbasResult}
\title{Convert a `dbasResult` to a `list` of `featureRecord`s}
\usage{
\method{convertToRecord}{dbasResult}(scanResult, msrawfileRecords)
}
\description{
The results of DBAS file scanning (`scanResult`) are converted to the NTSPortal `featureRecord` format.
Compound and sample metadata are collected from the `msrawfileRecords` argument and the spectral library (CSL).
}
\examples{
\dontrun{
dbComm <- getDbComm()
recs <- getTableAsRecords(
  dbComm, 
  "ntsp25.2_msrawfiles", 
  searchBlock = list(query = list(regexp = list(filename = "Des_.._.._pos.mzXML"))), 
  newMsrawfilesRecord
)
recsBlanks <- getTableAsRecords(
  dbComm, 
  "ntsp25.2_msrawfiles", 
  searchBlock = list(query = list(regexp = list(path = ".*mud_pos/BW.*"))), 
  newMsrawfilesRecord
)
res <- scanBatchDbas(c(recs, recsBlanks), "Methyltriphenylphosphonium")
featureRecs <- convertToRecord(res, c(recs, recsBlanks))
}
}
