% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{create_dbas_index}
\alias{create_dbas_index}
\title{Create a new, empty dbas index}
\usage{
create_dbas_index(escon, rfindex, aliasName, dateNum = NULL)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{aliasName}{Name of alias used in elasticsearch}

\item{dateNum}{Provide a specific version code (YYMMDD), default is NULL (current date)}
}
\value{
Invisibly returns index name of the created index
}
\description{
Will create a new index based on the alias with _v<date> appended after _dbas.
The new index name will be written to msrawfiles index.
}
