% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{createAllStructures}
\alias{createAllStructures}
\title{Create images of structural formulas for display in Kibana dashboard}
\usage{
createAllStructures(databasePath, targetDir)
}
\arguments{
\item{databasePath}{Path to spectral library (sqlite)}

\item{targetDir}{Path where the PNGs are stored}
}
\value{
no return value
}
\description{
Create images of structural formulas for display in Kibana dashboard
}
