% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningOneBatch}
\alias{dbaScreeningOneBatch}
\title{Process one batch of msrawfile records}
\usage{
dbaScreeningOneBatch(msrawfileRecords, saveDirectory)
}
\description{
Process one batch of msrawfile records
}
