% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-launchProcess.R
\name{dbaScreeningOneBatch}
\alias{dbaScreeningOneBatch}
\title{Process one batch of records}
\usage{
dbaScreeningOneBatch(msrawfileRecords, saveDirectory)
}
\description{
A function to be used only by the workload manager SLURM.
}
