% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R,
%   R/dbInterface-PythonDbComm.R
\docType{methods}
\name{isTable}
\alias{isTable}
\alias{getTableAsRecords}
\alias{getTableAsTibble}
\alias{setValueInField}
\alias{ping}
\alias{ping,PythonDbComm-method}
\alias{show,PythonDbComm-method}
\title{Does a table exist?}
\usage{
isTable(dbComm, tableName)

getTableAsRecords(
  dbComm,
  tableName,
  searchBlock = list(),
  recordConstructor = newNtspRecord
)

getTableAsTibble(dbComm, tableName, searchBlock = list())

setValueInField(dbComm, tableName, field, value, searchBlock = list())

ping(dbComm)

\S4method{ping}{PythonDbComm}(dbComm)

\S4method{show}{PythonDbComm}(object)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database}

\item{searchBlock}{list coercible to json for elasticSearch REST API (Query DSL)}

\item{recordConstructor}{function to construct new records}

\item{field}{table column to change}

\item{value}{value to set in field}
}
\value{
logical, TRUE if table exists

A list of ntspRecord objects

A tibble

TRUE when connection active
}
\description{
Does a table exist?

Get a table as a list of ntspRecords

Get a table as a tibble

Set the value of a field in a table

Test the DB-Connection
}
