% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-PythonDbComm.R
\docType{methods}
\name{getTableAsRecords}
\alias{getTableAsRecords}
\alias{ping}
\alias{show,PythonDbComm-method}
\alias{ping,PythonDbComm-method}
\title{Test the DB-Connection}
\usage{
getTableAsRecords(
  dbComm,
  tableName,
  queryBlock = list(),
  recordConstructor = newNtspRecord
)

ping(dbComm)

\S4method{show}{PythonDbComm}(object)

\S4method{ping}{PythonDbComm}(dbComm)
}
\arguments{
\item{queryBlock}{list coercible to json for elasticSearch REST API (Query DSL)}

\item{recordConstructor}{function to construct new records}
}
\value{
TRUE when connection active

TRUE when connection active
}
\description{
Test the DB-Connection

Test the DB-Connection
}
