% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_field}
\alias{es_add_field}
\title{Add a value to any field}
\usage{
es_add_field(escon, esindex, field, queryBody, value)
}
\arguments{
\item{escon}{}

\item{esindex}{}

\item{field}{}

\item{queryBody}{query in list format}

\item{value}{}
}
\description{
Only works for length 1 values
}
