% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_rt_cluster}
\alias{es_add_rt_cluster}
\title{Add rt_clustering field to es index pattern}
\usage{
es_add_rt_cluster(escon, esindex)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{esindex}{Index or index-pattern to update}
}
\value{
Returns the number of documents updated, invisibly
}
\description{
Add rt_clustering field to es index pattern
}
