% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_rtt}
\alias{es_add_rtt}
\title{Add rtt field to all documents where this does not exist}
\usage{
es_add_rtt(escon, esindex)
}
\arguments{
\item{escon}{connection to ElasticSearch}

\item{esindex}{Index or index-pattern to update}
}
\value{
Number of updated docs (invisibly)
}
\description{
Add rtt field to all documents where this does not exist
}
