% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_add_ufid}
\alias{es_add_ufid}
\title{Update doc in ntsportal with ufid}
\usage{
es_add_ufid(feat, escon, index)
}
\arguments{
\item{feat}{ntsportal::feature object}

\item{escon}{elasticsearch connection object created by elastic::connect}

\item{index}{elasticsearch index name}
}
\value{
TRUE if update successful
}
\description{
all the information needed is
contained within the feature object
}
