% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_add_ufid_to_ids}
\alias{es_add_ufid_to_ids}
\title{Add ufid entry to a list of features in elasticsearch}
\usage{
es_add_ufid_to_ids(escon, index, ufid_to_add, ids_for_update)
}
\arguments{
\item{ids_for_update}{}
}
\value{
TRUE if successful
}
\description{
Add ufid entry to a list of features in elasticsearch
}
