% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_data_to_docs.R
\name{es_add_value}
\alias{es_add_value}
\title{Add a value to any field in ES index}
\usage{
es_add_value(escon, esindex, field, queryBody, value)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{esindex}{Index name or index-pattern}

\item{field}{Field name to which value is added, field must already exist.}

\item{queryBody}{query in list format}

\item{value}{Value to add (must be length 1)}
}
\value{
API response (list)
}
\description{
Only works for length 1 values
}
