% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_assign_ufids}
\alias{es_assign_ufids}
\title{Go sequentially through ufid-db and assign ufids in es-db}
\usage{
es_assign_ufids(escon, udb, index, polarity, ufids = NULL)
}
\arguments{
\item{ufids}{if ufids is NULL (default), then all ufids will be processed}
}
\value{

}
\description{
updates ufid-db as well when it is done. This can either be done for all
ufids, or for a selection by giving a vector of integers.
}
