% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_feat_from_id}
\alias{es_feat_from_id}
\title{Get feature (doc) from ntsp with doc-id}
\usage{
es_feat_from_id(escon, index, id)
}
\arguments{
\item{id}{}
}
\description{
Get feature (doc) from ntsp with doc-id
}
