% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{es_move_alias}
\alias{es_move_alias}
\title{Change alias of an index to a new index.}
\usage{
es_move_alias(escon, indexName, aliasName)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{indexName}{}

\item{aliasName}{}
}
\description{
Will delete the previous alias
}
