% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{es_remove_by_filename}
\alias{es_remove_by_filename}
\title{Removal of documents from dbas documents based on filename}
\usage{
es_remove_by_filename(escon, index, filenames)
}
\arguments{
\item{escon}{}

\item{index}{}

\item{filenames}{}
}
\description{
Removal of documents from dbas documents based on filename
}
