% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_db_comm.R
\name{es_remove_ufid}
\alias{es_remove_ufid}
\title{Delete all ufid entries for a ufid}
\usage{
es_remove_ufid(escon, index, ufidToDelete)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{index}{Elasticsearch index name}

\item{ufidToDelete}{ID of ufid to delete}
}
\value{
TRUE invisibly
}
\description{
Delete all ufid entries for a ufid
}
