% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_testing.R
\name{es_test_fpfn}
\alias{es_test_fpfn}
\title{Compute validation statistics for ufid and ucid assignment}
\usage{
es_test_fpfn(escon, index, includeUfid2 = FALSE, includeUcid = FALSE)
}
\arguments{
\item{includeUfid2}{include ufid2 stats}

\item{includeUcid}{include ucid stats}
}
\value{
list of metrics used to assess quality of ufid and ucid assignments
}
\description{
Used for testing purposes.
}
