% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{get_field}
\alias{get_field}
\title{Get a field for a set of documents}
\usage{
get_field(esids, fieldName, simplify = T, justone = F)
}
\arguments{
\item{esids}{}

\item{fieldName}{}

\item{simplify}{}

\item{justone}{}
}
\description{
Get a field for a set of documents
}
