% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-fileScanning.R
\name{get_field}
\alias{get_field}
\title{Get a field for a set of documents defined by id}
\usage{
get_field(
  escon,
  indexName,
  esids = NULL,
  query = NULL,
  fieldName,
  simplify = T,
  justone = F
)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{esids}{Document IDs in the named index, must provide either esids or query}

\item{query}{Query to get docs in the form of a list, must provide either esids or query}

\item{fieldName}{field to extract}

\item{simplify}{Boolean, true will try to create a vector from the results}

\item{justone}{Boolean. If set to true, function will return only one result and will hang if cardinality > 1}

\item{index}{Name Elasticsearch index name}
}
\value{
Vector or list of the field entries, returns NA is nothing found
}
\description{
Get a field for a set of documents defined by id
}
