% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-fileScanning.R
\name{get_field_builder}
\alias{get_field_builder}
\title{Function factory just to avoid typing escon and index every time}
\usage{
get_field_builder(escon, index)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Elasticsearch index name}
}
\value{
get_field function with escon and index arguments set as default
}
\description{
Function factory just to avoid typing escon and index every time
}
