% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R, R/utilities.R
\name{get_field_builder}
\alias{get_field_builder}
\title{Function factory just to avoid typing escon and index every time}
\usage{
get_field_builder(escon, index)

get_field_builder(escon, index)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{index}{Elasticsearch index name}
}
\value{
get_field function with escon and index arguments set as default

get_field function with default ca
}
\description{
Function factory just to avoid typing escon and index every time

Create a get_field function with some default values
}
