% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_next_ufid}
\alias{get_next_ufid}
\title{Return a new, unique ufid number}
\usage{
get_next_ufid(udb, escon, index)
}
\arguments{
\item{udb}{}

\item{escon}{}

\item{index}{}
}
\value{
length one integer
}
\description{
The next ufid number is returned, for creating a new cluster. Will double
check that this number is not already present.
}
