% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{get_rare}
\alias{get_rare}
\title{Get rare compounds from a Report}
\usage{
get_rare(repo, min_freq)
}
\arguments{
\item{repo}{}

\item{min_freq}{}
}
\description{
Get the compounds which are found fewer than min_freq
}
