% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recall.R
\name{get_time_series_known}
\alias{get_time_series_known}
\title{Get entire time series for one compound at one station}
\usage{
get_time_series_known(
  escon,
  index = "g2_dbas_bfg",
  compound = "Carbamazepine",
  station = "rhein_ko_l",
  pol = "pos",
  matrix = "water",
  calendarInterval = "day"
)
}
\arguments{
\item{escon}{Elastic connection opject}

\item{index}{Index name}

\item{compound}{Compound name to filter data by (see ntsportal for correct spelling)}

\item{station}{Station name to filter data by (see available station codes in the ntsportal wiki)}

\item{pol}{Polarity to filter data by (either "pos" or "neg")}

\item{matrix}{Matrix code to filter data by (see available matrix codes in the ntsportal wiki)}

\item{calendarInterval}{Calendar internval, see available options in elasticsearch documentations for `date_histogram`}
}
\description{
Function should be expanded to allow user to choose date range and get data
from more than one compound, more than one station.
}
