% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{getTableAsRecords}
\alias{getTableAsRecords}
\title{Get a table as a list of ntspRecords}
\usage{
getTableAsRecords(
  dbComm,
  tableName,
  searchBlock = list(),
  recordConstructor = newNtspRecord
)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}

\item{recordConstructor}{function to construct new records}
}
\value{
A list of ntspRecord objects
}
\description{
Get a table as a list of ntspRecords
}
