% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{getTableAsTibble}
\alias{getTableAsTibble}
\title{Get a table as a tibble}
\usage{
getTableAsTibble(dbComm, tableName, searchBlock = list(), fields = "*")
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}

\item{fields}{Fields to include in the response, wildcards are permitted, default is all fields.}
}
\value{
A tibble
}
\description{
Works similarly to `getTableAsRecords()` but reformats the `ntspRecords` into a `tibble`.
}
\seealso{
\link{`getTableAsRecords`}
}
