% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-ingest.R
\name{ingest}
\alias{ingest}
\title{Ingest of RDS files into Elasticsearch}
\usage{
ingest(path, ingestPipeline = NULL)
}
\arguments{
\item{path}{Path to single RDS file or directory with RDS files.}

\item{ingestPipeline}{ID of ingest pipeline to use (default is none used)}
}
\value{
list with one top-level entry per batch (RDS file). The second level has one entry per
alias and this contains a character vector of the index name associated with that alias.
}
\description{
The processing output is an RDS file for each batch (or more than one for large batches). These
are ingested into NTSPortal Elasticsearch.
}
\details{
\strong{Ingest pipelines}
Feature documents are enriched with sample and analytical method information using an ingest pipeline.
Use `ingestPipeline` to specify which ingest pipeline to use (must have been previously submitted to Elasticsearch)
}
