% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-ingest.R
\name{ingestJson}
\alias{ingestJson}
\title{Ingest of json files into Elasticsearch}
\usage{
ingestJson(json_path)
}
\arguments{
\item{json_path}{Path to single json file or directory with json files.}
}
\value{
Returns TRUE when completed (invisibly)
}
\description{
Ingest of json files into Elasticsearch
}
