% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-ingest.R
\name{ingestJson}
\alias{ingestJson}
\title{Ingest of json.gz files into Elasticsearch}
\usage{
ingestJson(path)
}
\arguments{
\item{path}{Path to single json file or directory with json files.}
}
\value{
Returns pairs for all unique aliases and the respective indeces
}
\description{
Ingest of json.gz files into Elasticsearch
}
