% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_generic.R
\name{make_ntspl}
\alias{make_ntspl}
\title{Convert processing results to a list structure for ntsportal}
\usage{
make_ntspl(proco, coresBatch = 1)
}
\arguments{
\item{proco}{Processing output of class proco}

\item{coresBatch}{number of cores to use for parallelization (at the moment only 1)}
}
\value{
a list object of class ntspl
}
\description{
Transforms a process output object (proco) which has a structure the workflow
used to a ntspl list object which has the structure for ntsportal
}
\details{
This is a generic with methods for different workflow outputs. The output
for ntsworkflow has the class proco_nts (see make_ntspl.proco_nts).
}
