% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msrawfiles-versionIncrementer.R
\name{msrawfilesSetVersion}
\alias{msrawfilesSetVersion}
\title{Change msrawfiles to a new version}
\usage{
msrawfilesSetVersion(msrawfilesName, version)
}
\arguments{
\item{msrawfilesName}{Name of msrawfiles table}

\item{version}{New version number}
}
\description{
Change msrawfiles to a new version
}
\examples{
\dontrun{
msrawfilesSetVersion("msrawfiles", "25.1")
}
}
