% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_class.R
\name{new_feature}
\alias{new_feature}
\title{Create feature}
\usage{
new_feature(
  pol = character(),
  mz = double(),
  rt = double(),
  es_id = character(),
  chrom_method = character(),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Create feature
}
