% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{norm_ms1}
\alias{norm_ms1}
\title{Normalize ms1 spectrum to the maximum intensity}
\usage{
norm_ms1(x, precursorMz, noiselevel = 0.01)
}
\arguments{
\item{x}{}

\item{noiselevel}{}

\item{precursormz}{}

\item{precursorInt}{}
}
\description{
Normalize ms1 spectrum to the maximum intensity
}
