% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{norm_ms2}
\alias{norm_ms2}
\title{Normalize ms2 spectrum to the maximum intensity}
\usage{
norm_ms2(x, precursorMz, noiselevel = 0.01)
}
\arguments{
\item{x}{}

\item{noiselevel}{}

\item{precursormz}{}

\item{mztol}{}
}
\description{
Normalize ms2 spectrum to the maximum intensity
}
