% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntsportal-package.R
\docType{package}
\name{ntsportal-package}
\alias{ntsportal-package}
\title{ntsportal: ntsportal - Functions for building and managing the NTSPortal database}
\description{
This package provides functionality for building and managing the NTSPortal database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bafg-bund/ntsportal}
}

}
\author{
\strong{Maintainer}: Kevin S. Jewell \email{jewell@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)

Authors:
\itemize{
  \item Ole Lessmann \email{lessmann@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
}

Other contributors:
\itemize{
  \item Jonas Skottnik \email{skottnik@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [contributor]
  \item Franziska Prodöhl \email{prodoehl@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [contributor]
  \item Nina Hermes \email{hermes@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [contributor]
  \item Arne Wick \email{wick@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany) [reviewer]
}

}
\keyword{internal}
