% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntsportal-package.R
\docType{package}
\name{ntsportal-package}
\alias{ntsportal-package}
\title{ntsportal: NTSPortal database backend. Federal Institute of Hydrology, Koblenz, Germany}
\description{
This package provides functionality for building and managing the NTSPortal database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bafg-bund.github.io/ntsportal}
}

}
\author{
\strong{Maintainer}: Kevin S. Jewell \email{jewell@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)

Authors:
\itemize{
  \item Ole Lessmann \email{lessmann@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Jonas Skottnik \email{skottnik@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Franziska Thron \email{thron@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Arne Wick \email{wick@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
}

}
\keyword{internal}
