% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{proc_batch}
\alias{proc_batch}
\title{Process a batch of MS measurement files stored in the msrawfiles index}
\usage{
proc_batch(
  escon,
  rfindex,
  esids,
  tempsavedir,
  ingestpth,
  configfile,
  coresBatch,
  noIngest = FALSE
)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{esids}{ElasticSearch document IDs (character)}

\item{tempsavedir}{Temporary save location}

\item{ingestpth}{Path where the ingest.sh script is found}

\item{configfile}{Config file where the credentials for signing into elasticsearch are found}

\item{coresBatch}{Number of cores to use in a signal batch}

\item{noIngest}{Logical, for testing purposes, no upload, just create json.}
}
\value{
Function is run for side-effects but returns number of documents uploaded
}
\description{
Takes a vector of esids and performs dbas on these as a batch. Files are
evaluated using the settings in the msrawfiles index. Results are ingested to
the indicated dbas results index
}
