% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{proc_batch}
\alias{proc_batch}
\title{Process a batch of MS measurement files stored in the msrawfiles index}
\usage{
proc_batch(
  escon,
  rfindex,
  esids,
  tempsavedir,
  ingestpth,
  configfile,
  coresBatch,
  noIngest = FALSE
)
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{esids}{}

\item{tempsavedir}{}

\item{ingestpth}{}

\item{configfile}{}

\item{coresBatch}{}

\item{noIngest}{logical, for testing purposes, no upload, just create json.}
}
\description{
takes a series of esids and processes these as a batch.
}
