% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{proc_batch_nts}
\alias{proc_batch_nts}
\title{Process a batch of files using ntsworkflow}
\usage{
proc_batch_nts(
  escon,
  rfindex,
  esids,
  tempsavedir,
  ingestpth,
  configfile,
  coresBatch,
  noIngest = FALSE
)
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{rfindex}{Elasticsearch index name for msrawfiles index}

\item{esids}{ElasticSearch document IDs (character)}

\item{tempsavedir}{Temporary save location}

\item{ingestpth}{Path where the ingest.sh script is found}

\item{configfile}{Config file where the credentials for signing into elasticsearch are found}

\item{coresBatch}{Number of cores to use in a signal batch}

\item{noIngest}{Logical, for testing purposes, no upload, just create json.}
}
\description{
Still in development
}
