% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rawfiles_nts.R
\name{proc_batch_nts}
\alias{proc_batch_nts}
\title{Process a batch of files using ntsworkflow}
\usage{
proc_batch_nts(docsList, coresBatch)
}
\arguments{
\item{docsList}{ntsp_docs_list list of documents as returned by elasticSearch API (hits.hits array)}

\item{coresBatch}{Number of cores to use in a signal batch}
}
\value{
Output of processing. An object of class proco_nts
}
\description{
Will take a list of documents and process samples in them for 
peak-picking, alignment, blank correction, annotation
}
\details{
# Filtering the alignment table 
## Consecutive filter 
The order of the samples is important for the
consecutive filter, these must be sorted by start time prior to starting
this processing.
}
