% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{proc_esid}
\alias{proc_esid}
\title{Process an MS measurement file}
\usage{
proc_esid(escon, rfindex, esid)
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{esid}{}
}
\value{
an object of class ntsworkflow::Report
}
\description{
Takes an esid from the rawfiles index and carries out dbas processing for that file
}
