% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{proc_esid}
\alias{proc_esid}
\title{Process an MS measurement file}
\usage{
proc_esid(escon, rfindex, esid, compsProcess = NULL)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{esid}{ElasticSearch document IDs (character)}
}
\value{
an object of class ntsworkflow::Report
}
\description{
Takes an esid from the rawfiles index and carries out dbas processing for 
that file
}
