% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{proc_is_one}
\alias{proc_is_one}
\title{Process IS in one file}
\usage{
proc_is_one(escon, rfindex, esid)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{esid}{ID of document in rfindex}
}
\description{
Process IS in one file
}
