% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{process_is_all}
\alias{process_is_all}
\title{Process all files in msrawfiles for IS}
\usage{
process_is_all(
  escon,
  rfindex,
  isindex,
  ingestpth,
  configfile,
  tmpPath = "/scratch/nts/tmp",
  numCores = 10
)
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{isindex}{}

\item{ingestpth}{}

\item{configfile}{}

\item{tmpPath}{}

\item{numCores}{Number of cores for parallel processing}
}
\description{
Will enter the results into the isindex. Will only process files which
are not included in isindex.
}
