% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{process_is_all}
\alias{process_is_all}
\title{Process all files in msrawfiles for IS}
\usage{
process_is_all(
  escon,
  rfindex,
  isindex,
  ingestpth,
  configfile,
  tmpPath = "/scratch/nts/tmp",
  numCores = 10,
  rawfilesRootPath = "/scratch/nts/messdaten/",
  numFilesToProcess = NULL
)
}
\arguments{
\item{escon}{Elasticsearch connection object created by `elastic::connect`}

\item{rfindex}{Name of rawfiles index}

\item{isindex}{}

\item{ingestpth}{Path where the ingest.sh script is found}

\item{configfile}{Config file where the credentials for signing into elasticsearch are found}

\item{tmpPath}{Temporary save location}

\item{numCores}{Number of cores for parallel processing}

\item{rawfilesRootPath}{Path to where raw data is stored}

\item{numFilesToProcess}{If NULL (default) all remaining files are processed,
otherwise must be a positive integer.}
}
\description{
Will enter the results into the isindex. Will only process files which
are not included in isindex.
}
