% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_mappings.R
\name{put_dbas_index}
\alias{put_dbas_index}
\title{Send index creation request for dbas index}
\usage{
put_dbas_index(eson, index)
}
\arguments{
\item{eson}{elasticsearch connection object created with elastic::connect}

\item{index}{Name of the index you wish to create}
}
\description{
Must be kept in sync with: https://github.com/bafg-bund/ntsportal/wiki/Index-mapping-for-dbas-indices
}
