% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-PythonDbComm.R
\name{PythonDbComm}
\alias{PythonDbComm}
\title{Create connection to elasticsearch}
\usage{
PythonDbComm(ring = "ntsportal")
}
\arguments{
\item{ring}{}
}
\value{
PythonDbComm
}
\description{
Create connection to elasticsearch
}
