% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{rawfiles_missing}
\alias{rawfiles_missing}
\title{Check if any files are missing}
\usage{
rawfiles_missing(escon, rfindex)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}
}
\description{
Check if any files are missing
}
