% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructuralFormulas.R
\name{removeRedundantPngs}
\alias{removeRedundantPngs}
\title{Delete redundant PNGs}
\usage{
removeRedundantPngs(databasePath, targetDir)
}
\arguments{
\item{databasePath}{Path to spectral library (CSL, SQLite format)}

\item{targetDir}{Path where the PNGs are to be saved}
}
\description{
Remove PNGs for compounds not found in the library
}
