% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{replaceValueInField}
\alias{replaceValueInField}
\title{Replace a value in a field}
\usage{
replaceValueInField(dbComm, tableName, field, oldValue, newValue)
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{field}{table column (field) name see mappings}

\item{oldValue}{Value to be replaced}
}
\description{
Replace a value in a field
}
