% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rawfiles_dbas.R
\name{reset_eval}
\alias{reset_eval}
\title{Reset files in msrawfiles, so that they will be processed again.}
\usage{
reset_eval(escon, rfindex, queryBody, indexType = "dbas")
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{queryBody}{}

\item{indexType}{}
}
\description{
If there are errors it may be necessary to reprocess files. This function
will remove the fields "dbas_last_eval" and "dbas_spectral_library_sha256"
so that during the next processing, the files will be processed again.
}
