% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionToElasticsearch.R
\name{resetConnectionCredentials}
\alias{resetConnectionCredentials}
\title{Create connection with new credentials}
\usage{
resetConnectionCredentials()
}
\description{
Create connection with new credentials
}
