% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-fileScanning.R
\name{scanBatchDbas}
\alias{scanBatchDbas}
\title{Scan measurement files for known compounds}
\usage{
scanBatchDbas(records, compsToProcess = NULL, showProgress = FALSE)
}
\arguments{
\item{records}{`list` of `msrawfileRecord`s}

\item{compsToProcess}{Character vector of compounds names to include in the scanning (default is all compounds in the spectral library)}

\item{showProgress}{Logical, show progress bar? (for interacte use, default false)}
}
\value{
a `dbasResult` (`list` of 7 tables) including `peakList`, `reintegrationResults` etc.
}
\description{
The spectral library is used to scan mzXML measurment files for compounds using the DBAS algorithm.
The parameters for DBAS and file locations (measurement files and spectral library) are stored in the `msrawfileRecord`s 
passed to the `records` argument.
}
\examples{
\dontrun{
dbComm <- getDbComm()
recs <- getTableAsRecords(
  dbComm, 
  "ntsp25.2_msrawfiles", 
  searchBlock = list(query = list(regexp = list(filename = "Des_19_.._pos.mzXML"))),
  newMsrawfilesRecord
)
recsBlanks <- getTableAsRecords(
  dbComm, 
  "ntsp25.2_msrawfiles", 
  searchBlock = list(query = list(regexp = list(path = ".*mud_pos/BW.*"))), 
  newMsrawfilesRecord
)
res <- scanBatchDbas(c(recs, recsBlanks), "Methyltriphenylphosphonium")
}
}
