% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-DbComm.R
\name{setValueInField}
\alias{setValueInField}
\title{Set the value of a field in a table}
\usage{
setValueInField(dbComm, tableName, field, value, searchBlock = list())
}
\arguments{
\item{dbComm}{DbComm connection object}

\item{tableName}{Name of table in database, wildcards are permitted}

\item{field}{table column (field) name see mappings}

\item{value}{value to set or add}

\item{searchBlock}{list coercible to json for Elasticsearch Search API (Query DSL)}
}
\description{
Used to change the value of a field for all rows (i.e. docs) returned by `seearchBlock`
}
