% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_testing.R
\name{silhouette_score}
\alias{silhouette_score}
\title{Compute silhouette score for all ufids in the index}
\usage{
silhouette_score(escon, testIndex, ufidType = "ufid")
}
\arguments{
\item{escon}{Connection object created with `elastic::connect`}

\item{testIndex}{Elasticsearch index name}

\item{ufidType}{Either "ufid" or "ufid2"}
}
\value{
dataframe with the scores for each feature
}
\description{
Compute silhouette score for all ufids in the index
}
