% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{station_from_code}
\alias{station_from_code}
\title{Get station and location based on othe docs in msrawfiles index}
\usage{
station_from_code(escon, rfindex, filename, stationRegex)
}
\arguments{
\item{escon}{}

\item{rfindex}{}

\item{filename}{}

\item{stationRegex}{Must have brackets for stringr::str_match}
}
\value{
list with geolocation and station name
}
\description{
Get station and location based on othe docs in msrawfiles index
}
