% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rawfiles.R
\name{station_from_code}
\alias{station_from_code}
\title{Get station, river, and geopoint location based on other docs in msrawfiles index}
\usage{
station_from_code(escon, rfindex, filename, stationRegex)
}
\arguments{
\item{escon}{elastic connection object created by elastic::connect}

\item{rfindex}{index name for rawfiles index}

\item{filename}{File from which to extract the station code, this code is then used to find other files}

\item{stationRegex}{Regex to extract the station code. It must have brackets since it uses stringr::str_match}
}
\value{
list with geolocation and station name
}
\description{
Get station, river, and geopoint location based on other docs in msrawfiles index
}
