% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-new-ufid.R
\name{ubd_new_ufid}
\alias{ubd_new_ufid}
\title{Add new ufid entry by dbscan clustering}
\usage{
ubd_new_ufid(
  ubd,
  escon,
  index,
  polarity,
  minPoints1 = 5,
  minPoints2 = 2,
  mzPosition = 0,
  rtPosition = 0
)
}
\arguments{
\item{ubd}{ufid database connection object (DBI::dbConnect)}

\item{escon}{NTSPortal connection object (elastic::connect)}

\item{index}{Index name}

\item{polarity}{polarity of features to process}

\item{minPoints1}{minimum number of features for first stage clustering}

\item{minPoints2}{minimum number of features for second stage clustering}

\item{mzPosition}{minimum mz (Da) for returning new features from NTSPortal}

\item{rtPosition}{minimum RT (min) for returning new features from NTSPortal}
}
\value{
TRUE if the function ended successfully
}
\description{
This is the main function for the clustering stage of NTSP alignment.
New features, which do not yet have a ufid entry, are clustered based on 
their similarity (within tolerances) in the m/z, tR and MS2 domains. New 
ufids are added both to NTSP and ufid-db.
}
