% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_db_comm.R
\name{udb_add_feature}
\alias{udb_add_feature}
\title{Add feature to ufid db}
\usage{
udb_add_feature(feat, udb)
}
\arguments{
\item{udb}{}
}
\value{

}
\description{
if this is not found in the database
for now only for the case where ms1 and ms2 are present
}
