% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_db_comm.R
\name{udb_clean_spectra}
\alias{udb_clean_spectra}
\title{Clean spectra within ufid db}
\usage{
udb_clean_spectra(udb, ufid, msLevel)
}
\arguments{
\item{udb}{Connection to ufid db, created by `DBI::dbConnect`}

\item{ufid}{ID of ufid (numeric)}

\item{msLevel}{Which spectral type (either 'ms1' or 'ms2')}
}
\value{
TRUE (invisibly) when successful
}
\description{
Clean spectra within ufid db
}
