% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_matching.R
\name{udb_feature_match}
\alias{udb_feature_match}
\title{Find ufid for feature}
\usage{
udb_feature_match(
  ftt,
  rtt,
  ms2t,
  ft,
  mztol = 0.007,
  rttol = 1,
  ms2dpThresh = 50
)
}
\arguments{
\item{ftt}{Precollected feature table from ufid-db}

\item{rtt}{Precollected retention_time table from ufid-db}

\item{ms2t}{Precollected ms2 table from ufid-db}

\item{ft}{`feature` object}

\item{mztol}{m/z tolerance in Da}

\item{rttol}{Retention time tolerance}

\item{ms2dpThresh}{Threshold for dot-product comparison}
}
\value{
ntsportal::ufid_assignment object (ufid and level) Currently only level 1 is
possible. Returns ufid_assignment of 0, level 0 if no match is found.
}
\description{
Recieves a feature and returns its ufid_assignment object based on
information in the ufid database. Currently, the feature must have m/z,
rt and ms2. Tables are precollected from the SQLite ufid library as tibbles to increase speed and allow parallelizatation
}
