% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_matching.R
\name{udb_feature_to_ufid_match}
\alias{udb_feature_to_ufid_match}
\title{Find if feature and specified ufid entry are a match}
\usage{
udb_feature_to_ufid_match(
  ftt,
  rtt,
  ms2t,
  ft,
  ufid_to_match,
  mztol = 0.007,
  rttol = 1,
  ms2dpThresh = 50
)
}
\arguments{
\item{ms2dpThresh}{}
}
\value{
true if matching
}
\description{
Find if feature and specified ufid entry are a match
}
