% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid_db_comm.R
\name{udb_remove_ufid}
\alias{udb_remove_ufid}
\title{Remove a ufid from the ufid library}
\usage{
udb_remove_ufid(udb, ufid)
}
\arguments{
\item{udb}{Connection to ufid db, created by `DBI::dbConnect`}

\item{ufid}{ID of ufid (numeric)}
}
\value{
TRUE (invisibly) when successful
}
\description{
Remove a ufid from the ufid library
}
