% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufid2_alignment.R
\name{ufid2_alignment}
\alias{ufid2_alignment}
\title{Alignment by m/z and RT}
\usage{
ufid2_alignment(
  escon,
  index = "g2_nts*",
  rtTol = 0.3,
  mzTolmDa = 5,
  minPoints = 5
)
}
\arguments{
\item{escon}{}

\item{index}{}

\item{rtTol}{}

\item{mzTolmDa}{}

\item{minPoints}{}
}
\value{
TRUE
}
\description{
blah blah blah
}
