% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rawfiles_db.R
\name{update_alias_all}
\alias{update_alias_all}
\title{Update alias names}
\usage{
update_alias_all(escon, rfindex)
}
\arguments{
\item{escon}{}

\item{rfindex}{}
}
\description{
This function will update all the aliases in msrawfiles to use the current
indices found in msrawfiles
}
