% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInterface-ingestPipelines.R
\name{updateIngestPipelines}
\alias{updateIngestPipelines}
\title{Update Ingest Pipelines needed for ntsportal on the elasticsearch cluster}
\usage{
updateIngestPipelines()
}
\description{
Update Ingest Pipelines needed for ntsportal on the elasticsearch cluster
}
