% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-linearRegression.R
\name{updateLinearRegressionTable}
\alias{updateLinearRegressionTable}
\title{Update the "analysis_dbas" table for Time Series SPM dashboard}
\usage{
updateLinearRegressionTable(sourceTableName)
}
\arguments{
\item{sourceTableName}{Table in ntsportal which should be analyzed (only dbas_upb)}
}
\description{
The analysis table shows the slope of the linear regression model for each
compound at each station for the DBAS results of yearly composite 
samples of SPM (in the `dbas_upb` table). The NTSPortal admin runs this function after
processing to update the regression results.
}
