% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpectralLibrary.R
\name{updateSpectralLibrary}
\alias{updateSpectralLibrary}
\title{Update the CSL spectral library on ntsportal}
\usage{
updateSpectralLibrary(rfTableName, specLibTableName)
}
\arguments{
\item{rfTableName}{Name of the msrawfiles table}

\item{specLibTableName}{Name for the new spectral library table (old table will be overwritten)}
}
\description{
The `spectral_library` table is a copy of the CSL to use as a reference in the Spectral Library dashboard 
and Spectra and Chromatograms dashboard. This does not change the library used for DBAS processing. After
reprocessing with a new CSL version, this function is run to update the reference table.
}
