% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntspRecord-validation.R
\name{validateRecord}
\alias{validateRecord}
\alias{validateRecord.msrawfilesRecord}
\alias{validateRecord.featureRecord}
\title{Validate that a record is correctly formated}
\usage{
validateRecord(record)

\method{validateRecord}{msrawfilesRecord}(record)

\method{validateRecord}{featureRecord}(record)
}
\description{
Validate that a record is correctly formated
}
